\name{CrossSpeciesCombat.ordered}
\alias{CrossSpeciesCombat.ordered}
\docType{data}
\title{ 
Ordered samples across species for Combat.
}
\description{
Ordered samples across species for Combat.
}
\usage{data("CrossSpeciesCombat.ordered")}
\format{
  A data frame with 2 observations on the following 83 variables.
  \describe{
    \item{\code{G}}{a character vector}
    \item{\code{I.9811.tumor}}{a numeric vector}
    \item{\code{I.9812.IC2}}{a numeric vector}
    \item{\code{II.6413.IC2}}{a numeric vector}
    \item{\code{II.6415.tumor}}{a numeric vector}
    \item{\code{II.648.IC2}}{a numeric vector}
    \item{\code{II.649.tumor}}{a numeric vector}
    \item{\code{III.8016.IC2}}{a numeric vector}
    \item{\code{III.8017.tumor}}{a numeric vector}
    \item{\code{III.8023.IC2}}{a numeric vector}
    \item{\code{III.8024.tumor}}{a numeric vector}
    \item{\code{X100}}{a numeric vector}
    \item{\code{X102}}{a numeric vector}
    \item{\code{X112}}{a numeric vector}
    \item{\code{X114}}{a numeric vector}
    \item{\code{X115}}{a numeric vector}
    \item{\code{X116}}{a numeric vector}
    \item{\code{X117}}{a numeric vector}
    \item{\code{X118}}{a numeric vector}
    \item{\code{X119}}{a numeric vector}
    \item{\code{X120}}{a numeric vector}
    \item{\code{X124}}{a numeric vector}
    \item{\code{X126}}{a numeric vector}
    \item{\code{X127}}{a numeric vector}
    \item{\code{X128}}{a numeric vector}
    \item{\code{X129}}{a numeric vector}
    \item{\code{X130}}{a numeric vector}
    \item{\code{X131}}{a numeric vector}
    \item{\code{X132}}{a numeric vector}
    \item{\code{X133}}{a numeric vector}
    \item{\code{X134}}{a numeric vector}
    \item{\code{X135}}{a numeric vector}
    \item{\code{X136}}{a numeric vector}
    \item{\code{X137}}{a numeric vector}
    \item{\code{X138}}{a numeric vector}
    \item{\code{X33}}{a numeric vector}
    \item{\code{X34}}{a numeric vector}
    \item{\code{X35}}{a numeric vector}
    \item{\code{X36}}{a numeric vector}
    \item{\code{X37}}{a numeric vector}
    \item{\code{X38}}{a numeric vector}
    \item{\code{X39}}{a numeric vector}
    \item{\code{X40}}{a numeric vector}
    \item{\code{X41}}{a numeric vector}
    \item{\code{X42}}{a numeric vector}
    \item{\code{X44}}{a numeric vector}
    \item{\code{X46}}{a numeric vector}
    \item{\code{X47}}{a numeric vector}
    \item{\code{X48}}{a numeric vector}
    \item{\code{X49}}{a numeric vector}
    \item{\code{X50}}{a numeric vector}
    \item{\code{X51}}{a numeric vector}
    \item{\code{X52}}{a numeric vector}
    \item{\code{X54}}{a numeric vector}
    \item{\code{X55}}{a numeric vector}
    \item{\code{X56}}{a numeric vector}
    \item{\code{X57}}{a numeric vector}
    \item{\code{X58}}{a numeric vector}
    \item{\code{X59}}{a numeric vector}
    \item{\code{X60}}{a numeric vector}
    \item{\code{X61}}{a numeric vector}
    \item{\code{X62}}{a numeric vector}
    \item{\code{X64}}{a numeric vector}
    \item{\code{X65}}{a numeric vector}
    \item{\code{X67}}{a numeric vector}
    \item{\code{X68}}{a numeric vector}
    \item{\code{X69}}{a numeric vector}
    \item{\code{X72}}{a numeric vector}
    \item{\code{X73}}{a numeric vector}
    \item{\code{X74}}{a numeric vector}
    \item{\code{X75}}{a numeric vector}
    \item{\code{X84}}{a numeric vector}
    \item{\code{X86}}{a numeric vector}
    \item{\code{X89}}{a numeric vector}
    \item{\code{X90}}{a numeric vector}
    \item{\code{X92}}{a numeric vector}
    \item{\code{X93}}{a numeric vector}
    \item{\code{X94}}{a numeric vector}
    \item{\code{X95}}{a numeric vector}
    \item{\code{X96}}{a numeric vector}
    \item{\code{X97}}{a numeric vector}
    \item{\code{X98}}{a numeric vector}
    \item{\code{X99}}{a numeric vector}
  }
}
\details{
Ordered samples across species for Combat. 
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(CrossSpeciesCombat.ordered)
## maybe str(CrossSpeciesCombat.ordered) ; plot(CrossSpeciesCombat.ordered) ...
}
\keyword{datasets}
