\name{EnrichmentAnalysis.labels1}
\alias{EnrichmentAnalysis.labels1}
\docType{data}
\title{ The PanNETassinger-miR subtypes for enrichment analysis.
}
\description{
The PanNETassinger-miR subtypes for enrichment analysis.}
\usage{data("EnrichmentAnalysis.labels1")}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{Name}}{a character vector}
    \item{\code{miRNA.subtypes}}{a numeric vector}
    \item{\code{miRNmiR.cluster.1.subtypes}}{a character vector}
  }
}
\details{
The miRNA.subtypes for enrichment analysis.}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(EnrichmentAnalysis.labels1)
## maybe str(EnrichmentAnalysis.labels1) ; plot(EnrichmentAnalysis.labels1) ...
}
\keyword{datasets}
