\name{EnrichmentAnalysis.labels2}
\alias{EnrichmentAnalysis.labels2}
\docType{data}
\title{ The PanNETassinger-mRNA subtypes for enrichment analysis.
}
\description{
The PanNETassinger-mRNA subtypes for enrichment analysis.}
\usage{data("EnrichmentAnalysis.labels2")}
\format{
  A data frame with 36 observations on the following 3 variables.
  \describe{
    \item{\code{Name.1}}{a numeric vector}
    \item{\code{Name}}{a character vector}
    \item{\code{membership.ordered}}{a character vector}
  }
}
\details{
The gene expression subtypes for enrichment analysis.}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(EnrichmentAnalysis.labels2)
## maybe str(EnrichmentAnalysis.labels2) ; plot(EnrichmentAnalysis.labels2) ...
}
\keyword{datasets}
