\name{MLPvsother.Indata}
\alias{MLPvsother.Indata}
\docType{data}
\title{
Combined MLP-1 and MLP-2 subtypes vs other subtypes from core clinical PanNET gene expression dataset.
}
\description{Combined MLP-1 and MLP-2 subtypes vs other subtypes from core clinical PanNET gene expression dataset.

}
\usage{data("MLPvsother.Indata")}
\format{
  A data frame with 2102 observations on the following 87 variables.
  \describe{
    \item{\code{Gene}}{a factor with levels \code{03-Sep} \code{A1CF} \code{A2BP1} \code{AAA1} \code{AADAC} \code{AANAT} \code{ABAT} \code{ABCA12} \code{ABCB11} \code{ABCB4} \code{ABCC5} \code{ABCC6} \code{ABCC8} \code{ABCD2} \code{ABCG2} \code{ABHD14A} \code{ABI3BP} \code{ABO} \code{ABP1} \code{ACAD9} \code{ACADSB} \code{ACE} \code{ACP5} \code{ACSF2} \code{ACSL1} \code{ACSM1} \code{ACSS3} \code{ACTN4} \code{ACTR3} \code{ACVR1B} \code{ADAM19} \code{ADAM2} \code{ADAM28} \code{ADAM8} \code{ADAMDEC1} \code{ADAMTS18} \code{ADAMTS2} \code{ADAMTS4} \code{ADAMTS7} \code{ADAMTSL2} \code{ADARB1} \code{ADCY2} \code{ADCY7} \code{ADH1A} \code{ADH4} \code{ADM} \code{ADRB1} \code{ADRM1} \code{AFF2} \code{AFF3} \code{AFG3L1} \code{AFM} \code{AGPAT2} \code{AGPAT4} \code{AGR2} \code{AGRP} \code{AGT} \code{AIM1} \code{AJAP1} \code{AK5} \code{AKAP12} \code{AKAP7} \code{AKR1B10} \code{AKR1C1} \code{AKR1C3} \code{AKR1C4} \code{ALB} \code{ALCAM} \code{ALDH1A1} \code{ALDH1A3} \code{ALDH2} \code{ALDH6A1} \code{ALDH9A1} \code{ALDOB} \code{ALDOC} \code{ALOX5} \code{ALX1} \code{AMICA1} \code{AMY2A} \code{ANAPC16} \code{ANGPT2} \code{ANGPTL1} \code{ANGPTL3} \code{ANGPTL4} \code{ANKMY2} \code{ANKRD11} \code{ANKRD13A} \code{ANKRD36B} \code{ANKS1B} \code{ANO3} \code{ANPEP} \code{ANXA11} \code{ANXA2} \code{AOAH} \code{AOX1} \code{AP1G2} \code{AP3B1} \code{APAF1} \code{APCDD1} \code{APLN} \code{APLP1} \code{APOA5} \code{APOBEC2} \code{APOC3} \code{APOD} \code{APOE} \code{APOH} \code{APOLD1} \code{APOM} \code{APP} \code{AQP1} \code{AQP3} \code{AQP8} \code{AQP9} \code{ARC} \code{AREG} \code{ARG1} \code{ARG2} \code{ARHGAP15} \code{ARHGAP6} \code{ARHGEF37} \code{ARHGEF4} \code{ARHGEF6} \code{ARHGEF7} \code{ARID1B} \code{ARID5B} \code{ARL17A} \code{ARRDC4} \code{ARSE} \code{ARSJ} \code{ART3} \code{ARTN} \code{AS3MT} \code{ASAM} \code{ASB4} \code{ASCL1} \code{ASGR1} \code{ASGR2} \code{ASH1L} \code{ASPH} \code{ASPHD2} \code{ASPN} \code{ASTN1} \code{ASTN2} \code{ASXL3} \code{ATF3} \code{ATF7IP2} \code{ATHL1} \code{ATP1B2} \code{ATP2A3} \code{ATP5C1} \code{ATP6V1B1} \code{ATP8A1} \code{ATP8A2} \code{B3GALT2} \code{B3GAT1} \code{B3GNT3} \code{B4GALT5} \code{B9D1} \code{BAALC} \code{BAAT} \code{BACE2} \code{BAGE} \code{BAHCC1} \code{BAIAP2L2} \code{BAIAP3} \code{BAMBI} \code{BANK1} \code{BASP1} \code{BATF3} \code{BAZ1A} \code{BBX} \code{BCAS1} \code{BCAT1} \code{BCHE} \code{BCL11A} \code{BEGAIN} \code{BEST1} \code{BFSP1} \code{BHLHE40} \code{BHLHE41} \code{BHMT} \code{BID} \code{BIK} \code{BLNK} \code{BMP5} \code{BMP7} \code{BNC2} \code{BOC} \code{BPIL1} \code{BRCA1} \code{BRCA2} \code{BSCL2} \code{BTBD19} \code{BTC} \code{BTG2} \code{BTG3} \code{BTNL8} \code{C10orf10} \code{C10orf116} \code{C10orf118} \code{C14orf105} \code{C14orf132} \code{C14orf139} \code{C14orf143} \code{C15orf48} \code{C15orf5} \code{C16orf7} \code{C17orf72} \code{C19orf77} \code{C1orf106} \code{C1orf113} \code{C1orf114} \code{C1orf116} \code{C1orf38} \code{C1orf95} \code{C1QA} \code{C1QTNF5} \code{C1S} \code{C20orf46} \code{C21orf62} \code{C2orf18} \code{C2orf27A} \code{C2orf34} \code{C2orf72} \code{C3orf32} \code{C3orf52} \code{C4A} \code{C4BPA} \code{C4orf31} \code{C4orf34} \code{C4orf7} \code{C5} \code{C5orf13} \code{C5orf23} \code{C6} \code{C6orf115} \code{C7orf10} \code{C7orf16} \code{C7orf68} \code{C8A} \code{C8B} \code{C8G} \code{C8orf79} \code{C9} \code{C9orf11} \code{C9orf116} \code{C9orf150} \code{C9orf3} \code{C9orf4} \code{C9orf45} \code{CA12} \code{CA2} \code{CA4} \code{CA5A} \code{CA9} \code{CABLES1} \code{CACNA1D} \code{CACNA2D1} \code{CACNA2D2} \code{CACNA2D3} \code{CACNA2D4} \code{CACNB4} \code{CADM1} \code{CADM3} \code{CADPS} \code{CADPS2} \code{CALB1} \code{CALB2} \code{CALCR} \code{CALCRL} \code{CALD1} \code{CALY} \code{CAMK1D} \code{CAMK2D} \code{CAMK2N1} \code{CAPN13} \code{CAPN9} \code{CAPNS1} \code{CAPZB} \code{CARHSP1} \code{CARTPT} \code{CASP1} \code{CASR} \code{CAT} \code{CBLC} \code{CBLN1} \code{CBR4} \code{CBX2} \code{CBX4} \code{CCBE1} \code{CCDC109B} \code{CCDC114} \code{CCDC3} \code{CCDC34} \code{CCDC68} \code{CCDC74B} \code{CCK} \code{CCKAR} \code{CCL11} \code{CCL13} \code{CCL15} \code{CCL18} \code{CCL19} \code{CCL2} \code{CCL20} \code{CCL21} \code{CCL24} \code{CCL26} \code{CCL3} \code{CCL4} \code{CCL8} \code{CCNA1} \code{CCND2} \code{CCNJL} \code{CCNO} \code{CCR1} \code{CCR10} \code{CCR7} \code{CCRL2} \code{CD109} \code{CD163} \code{CD27} \code{CD36} \code{CD44} \code{CD52} \code{CD82} \code{CDC42} \code{CDC45} \code{CDH11} \code{CDH16} \code{CDH17} \code{CDH18} \code{CDH2} \code{CDH22} \code{CDH4} \code{CDH6} \code{CDH7} \code{CDHR2} \code{CDHR5} \code{CDK1} \code{CDK5RAP3} \code{CDK6} \code{CDKN2C} \code{CDO1} \code{CDS1} \code{CDX2} \code{CEACAM1} \code{CEACAM5} \code{CEACAM7} \code{CEBPA} \code{CEBPD} \code{CECR1} \code{CELA1} \code{CELA3B} \code{CELF4} \code{CELF5} \code{CEP68} \code{CES4} \code{CETP} \code{CFC1} \code{CFI} \code{CGA} \code{CGN} \code{CH25H} \code{CHAD} \code{CHAT} \code{CHCHD3} \code{CHD5} \code{CHDH} \code{CHI3L2} \code{CHMP4C} \code{CHN2} \code{CHORDC1} \code{CHRNA3} \code{CHRNA5} \code{CHRNA7} \code{CHRNB3} \code{CHST1} \code{CHST11} \code{CHST4} \code{CHST8} \code{CKMT1A} \code{CKS2} \code{CLC} \code{CLCA1} \code{CLCN3} \code{CLCN4} \code{CLDN1} \code{CLDN10} \code{CLDN11} \code{CLDN15} \code{CLDN18} \code{CLDN2} \code{CLDN4} \code{CLDN7} \code{CLEC10A} \code{CLEC14A} \code{CLEC2B} \code{CLEC3A} \code{CLEC3B} \code{CLGN} \code{CLIC1} \code{CLIC4} \code{CLN6} \code{CLPS} \code{CLU} \code{CMYA5} \code{CNIH3} \code{CNN1} \code{CNPY2} \code{CNTFR} \code{CNTN1} \code{CNTN3} \code{CNTNAP2} \code{CNTNAP3B} \code{COCH} \code{COL10A1} \code{COL12A1} \code{COL14A1} \code{COL15A1} \code{COL16A1} \code{COL1A2} \code{COL27A1} \code{COL3A1} \code{COL4A1} \code{COL4A4} \code{COL5A2} \code{COL6A6} \code{COL7A1} \code{COL8A1} \code{COL9A1} \code{COMP} \code{COPE} \code{COPS8} \code{CORO2A} \code{CORO2B} \code{CORT} \code{COTL1} \code{COX4NB} \code{COX7A2} \code{CP} \code{CPA1} \code{CPA2} \code{CPA3} \code{CPB1} \code{CPB2} \code{CPE} \code{CPEB1} \code{CPLX2} \code{CPM} \code{CPN1} \code{CPNE7} \code{CPS1} \code{CPVL} \code{CRABP1} \code{CREB5} \code{CRH} \code{CRHR1} \code{CRIP1} \code{CRISP2} \code{CRISP3} \code{CRYAB} \code{CRYBA2} \code{CRYBB2} \code{CRYM} \code{CSDA} \code{CSRNP3} \code{CST4} \code{CSTA} \code{CTAGE5} \code{CTNNA2} \code{CTNND2} \code{CTRB1} \code{CTRC} \code{CTRL} \code{CTSE} \code{CTSK} \code{CTSL1} \code{CTSL2} \code{CTSS} \code{CTSW} \code{CTSZ} \code{CUX2} \code{CUZD1} \code{CX3CL1} \code{CX3CR1} \code{CXADR} \code{CXCL10} \code{CXCL11} \code{CXCL13} \code{CXCL14} \code{CXCL16} \code{CXCL2} \code{CXCL3} \code{CXCL6} \code{CXCR2} \code{CXCR4} \code{CXCR7} \code{CXorf57} \code{CXXC5} \code{CYB5R1} \code{CYBRD1} \code{CYP1A1} \code{CYP1B1} \code{CYP27A1} \code{CYP2A7} \code{CYP4F12} \code{CYP4F3} \code{CYTH3} \code{CYTH4} \code{CYTL1} \code{D4S234E} \code{DAAM1} \code{DACH1} \code{DACT1} \code{DAPK2} \code{DAPL1} \code{DAZ2} \code{DBC1} \code{DBH} \code{DCK} \code{DCLRE1C} \code{DCTN2} \code{DDC} \code{DDX25} \code{DEFA6} \code{DEFB1} \code{DGCR2} \code{DGCR8} \code{DHRS2} \code{DIRAS2} \code{DIRAS3} \code{DISC1} \code{DKK1} \code{DKK3} \code{DKK4} \code{DLG1} \code{DLK1} \code{DLL1} \code{DLL3} \code{DLL4} \code{DLX4} \code{DMBT1} \code{DMGDH} \code{DMKN} \code{DMRT2} \code{DMRTA1} \code{DNAJC12} \code{DNASE1} \code{DNASE1L3} \code{DNER} \code{DOCK7} \code{DOK5} \code{DOM3Z} \code{DPEP1} \code{DPP10} \code{DPP6} \code{DPT} \code{DPYD} \code{DPYSL3} \code{DPYSL4} \code{DSCAM} \code{DSCR6} \code{DSCR8} \code{DTNA} \code{DUSP1} \code{DYNC1I1} \code{DYRK1A} \code{DZIP1} \code{DZIP3} \code{ECE2} \code{ECEL1} \code{ECHDC3} \code{ECM1} \code{EDIL3} \code{EDN3} \code{EEF1A2} \code{EFEMP1} \code{EFEMP2} \code{EFHC2} \code{EFNA1} \code{EFNB3} \code{EFS} \code{EGFL6} \code{EGFL7} \code{EGFR} \code{EGLN3} \code{EGR1} \code{EGR2} \code{EGR3} \code{EGR4} \code{EHF} \code{EIF1AY} \code{EIF4G3} \code{ELAVL3} \code{ELAVL4} \code{ELF3} \code{ELMO1} \code{ELMOD1} \code{ELOVL2} \code{ELSPBP1} \code{EMCN} \code{EMP1} \code{ENC1} \code{ENG} \code{ENPP1} \code{ENPP2} \code{ENTPD3} \code{EPB41L4B} \code{EPHA4} \code{EPHA8} \code{EPHB3} \code{EPHX2} \code{EPS8L3} \code{ERAP2} \code{ERBB3} \code{ERBB4} \code{ERO1LB} \code{ERRFI1} \code{ESD} \code{ESM1} \code{ESPN} \code{ESRRG} \code{ETS1} \code{ETS2} \code{ETV1} \code{EVI2A} \code{EXD2} \code{EXOSC3} \code{EXT1} \code{F10} \code{F12} \code{F13A1} \code{F2RL1} \code{F7} \code{F9} \code{FABP3} \code{FABP4} \code{FABP5} \code{FABP6} \code{FAM101B} \code{FAM103A1} \code{FAM105A} \code{FAM107B} \code{FAM110B} \code{FAM133B} \code{FAM13A} \code{FAM13C} \code{FAM149A} \code{FAM167A} \code{FAM171A1} \code{FAM176A} \code{FAM198B} \code{FAM19A5} \code{FAM20A} \code{FAM20C} \code{FAM26F} \code{FAM3A} \code{FAM46C} \code{FAM49A} \code{FAM50B} \code{FAM59A} \code{FAM5B} \code{FAM63A} \code{FAM70A} \code{FAM75A7} \code{FAM83E} \code{FAM84A} \code{FAM90A1} \code{FAP} \code{FAT2} \code{FBLN1} \code{FBP1} \code{FBXL7} \code{FBXO2} \code{FCER1G} \code{FCGBP} \code{FCGR1A} \code{FCN3} \code{FDX1} \code{FEV} \code{FGB} \code{FGD1} \code{FGF12} \code{FGF13} \code{FGF14} \code{FGF17} \code{FGFR3} \code{FGG} \code{FGR} \code{FHOD3} \code{FICD} \code{FJX1} \code{FKBP11} \code{FKBP1B} \code{FLG} \code{FLJ13197} \code{FLJ20444} \code{FLJ39739} \code{FLJ43663} \code{FLNC} \code{FLRT1} \code{FLRT3} \code{FLT1} \code{FMN2} \code{FMNL1} \code{FMNL2} \code{FMOD} \code{FOLH1} \code{FOLR3} \code{FOSL2} \code{FOXO4} \code{FPR1} \code{FRMPD4} \code{FRZB} \code{FST} \code{FSTL4} \code{FSTL5} \code{FTH1} \code{FUS} \code{FXYD1} \code{FXYD3} \code{FYB} \code{FZD10} \code{FZD4} \code{FZD7} \code{GAB2} \code{GABRA2} \code{GABRB1} \code{GABRB3} \code{GABRG2} \code{GAD1} \code{GADD45B} \code{GADD45G} \code{GAGE2C} \code{GAL3ST1} \code{GAL3ST4} \code{GALNT10} \code{GALNT12} \code{GALNT4} \code{GALR1} \code{GAP43} \code{GAS2} \code{GAST} \code{GATA4} \code{GATA6} \code{GATM} \code{GBA2} \code{GBA3} \code{GC} \code{GCG} \code{GCGR} \code{GCHFR} \code{GCK} \code{GCLM} \code{GCNT3} \code{GDA} \code{GDF15} \code{GEM} \code{GFI1} \code{GFPT2} \code{GGH} \code{GHR} \code{GHRHR} \code{GHRL} \code{GIMAP6} \code{GJB2} \code{GLDC} \code{GLI1} \code{GLP1R} \code{GLRA2} \code{GLRB} \code{GLRX} \code{GLS} \code{GLT25D2} \code{GMDS} \code{GMFG} \code{GNAL} \code{GNAS} \code{GNAZ} \code{GNB1} \code{GNB3} \code{GNG4} \code{GP2} \code{GP5} \code{GPC3} \code{GPC4} \code{GPC5} \code{GPC6} \code{GPER} \code{GPI} \code{GPM6A} \code{GPNMB} \code{GPR12} \code{GPR133} \code{GPR158} \code{GPR160} \code{GPR19} \code{GPR34} \code{GPR37} \code{GPR44} \code{GPR64} \code{GPR83} \code{GPR85} \code{GPR88} \code{GPRASP1} \code{GPX2} \code{GPX3} \code{GRAMD2} \code{GRAMD3} \code{GRAMD4} \code{GRB14} \code{GREM1} \code{GRIA2} \code{GRIA3} \code{GRIA4} \code{GRID2} \code{GRIK4} \code{GRIN2C} \code{GRM1} \code{GRM5} \code{GRP} \code{GRSF1} \code{GSC2} \code{GSTA2} \code{GSTA4} \code{GSTP1} \code{GSTT1} \code{GTF2A2} \code{GUCA1C} \code{GUCY2C} \code{GULP1} \code{GYG2} \code{GYS1} \code{GZMA} \code{H2AFJ} \code{HAMP} \code{HAO1} \code{HAPLN1} \code{HAS2} \code{HBA1} \code{HBEGF} \code{HCG9} \code{HCLS1} \code{HDAC9} \code{HDC} \code{HECW2} \code{HEPH} \code{HERC4} \code{HERC5} \code{HES6} \code{HGD} \code{HHATL} \code{HIGD1B} \code{HIST1H1C} \code{HIST1H2BE} \code{HIST1H2BK} \code{HIST1H3H} \code{HIST1H4K} \code{HIST2H2AA3} \code{HIST2H4A} \code{HLA-DPA1} \code{HMBOX1} \code{HMGB2} \code{HMGCLL1} \code{HMGCS2} \code{HMMR} \code{HMOX1} \code{HMP19} \code{HNF4G} \code{HNRNPA1} \code{HNRNPH1} \code{HOPX} \code{HOXA10} \code{HOXA7} \code{HOXB13} \code{HOXB2} \code{HOXB4} \code{HOXB5} \code{HOXC11} \code{HOXC9} \code{HPD} \code{HPGD} \code{HPN} \code{HPX} \code{HR} \code{HRASLS} \code{HRG} \code{HS3ST1} \code{HS3ST2} \code{HS3ST3B1} \code{HS3ST4} \code{HS6ST2} \code{HSD11B1} \code{HSD11B2} \code{HSD17B2} \code{HSD17B6} \code{HSPA1B} \code{HSPA6} \code{HSPB1} \code{HSPD1} \code{HTR2B} \code{HYAL1} \code{IAPP} \code{ICAM1} \code{ICAM2} \code{ID1} \code{ID3} \code{ID4} \code{IDO1} \code{IDS} \code{IER2} \code{IER3} \code{IFFO2} \code{IFI30} \code{IFI44} \code{IFI6} \code{IFIT3} \code{IFNA8} \code{IGDCC4} \code{IGF2AS} \code{IGFBP1} \code{IGFBP2} \code{IGFBP3} \code{IGFBP5} \code{IGFBP6} \code{IGLC2} \code{IGSF1} \code{IL10RA} \code{IL11} \code{IL12A} \code{IL12RB2} \code{IL13RA2} \code{IL17RB} \code{IL18R1} \code{IL1B} \code{IL1R1} \code{IL1R2} \code{IL1RAP} \code{IL1RL1} \code{IL20RA} \code{IL22RA1} \code{IL33} \code{IL6} \code{IL6R} \code{IL7} \code{IL7R} \code{IL8} \code{IL9R} \code{IMPA2} \code{INCENP} \code{INHA} \code{INHBA} \code{INHBB} \code{INMT} \code{INPP5F} \code{INS} \code{INSL4} \code{INSM1} \code{IP6K2} \code{IRF6} \code{IRS1} \code{ISL1} \code{ISLR} \code{ISX} \code{ITGA10} \code{ITGA2} \code{ITGA7} \code{ITGB1BP1} \code{ITGB1BP2} \code{ITGB8} \code{ITGBL1} \code{ITIH2} \code{ITIH4} \code{ITM2A} \code{ITM2C} \code{ITPKC} \code{ITPR1} \code{ITPR3} \code{IWS1} \code{JAKMIP2} \code{KANK1} \code{KAZ} \code{KBTBD10} \code{KBTBD11} \code{KCNA1} \code{KCNA5} \code{KCNB1} \code{KCND2} \code{KCNH7} \code{KCNIP1} \code{KCNJ14} \code{KCNJ15} \code{KCNJ2} \code{KCNJ3} \code{KCNJ6} \code{KCNJ8} \code{KCNK1} \code{KCNK10} \code{KCNK3} \code{KCNK7} \code{KCNMA1} \code{KCNMB2} \code{KCNN2} \code{KCNQ2} \code{KCTD12} \code{KDR} \code{KIAA0141} \code{KIAA0319} \code{KIAA0408} \code{KIAA0649} \code{KIAA1107} \code{KIAA1211} \code{KIAA1239} \code{KIAA1244} \code{KIAA1310} \code{KIAA1324} \code{KIAA1370} \code{KIAA1549} \code{KIF17} \code{KIT} \code{KL} \code{KLF1} \code{KLF6} \code{KLHDC8A} \code{KLHL12} \code{KLHL26} \code{KLHL3} \code{KLK12} \code{KLK4} \code{KLK7} \code{KLK8} \code{KLKB1} \code{KLRB1} \code{KRT18} \code{KRT19} \code{KRT20} \code{KRT75} \code{KRT80} \code{LAG3} \code{LAIR1} \code{LAMA2} \code{LAMC2} \code{LAMC3} \code{LANCL2} \code{LAT2} \code{LCN2} \code{LDB2} \code{LDHB} \code{LEF1} \code{LEMD1} \code{LEPR} \code{LGALS2} \code{LGALS4} \code{LGI1} \code{LGI2} \code{LGR4} \code{LHFP} \code{LHX3} \code{LHX5} \code{LIF} \code{LIMCH1} \code{LIMS1} \code{LIMS2} \code{LIN7B} \code{LIPA} \code{LIPC} \code{LMCD1} \code{LMO1} \code{LMO3} \code{LNPEP} \code{LOC153684} \code{LOC154761} \code{LOC158863} \code{LOC253039} \code{LOC253264} \code{LOC256021} \code{LOC283070} \code{LOC286434} \code{LOC360030} \code{LOC388279} \code{LOC388458} \code{LOC388889} \code{LOC389831} \code{LOC399959} \code{LOC404266} \code{LOC439949} \code{LOC81691} \code{LOXL4} \code{LPAR1} \code{LPHN2} \code{LPIN1} \code{LPL} \code{LPPR1} \code{LPPR5} \code{LRAT} \code{LRCH2} \code{LRP5} \code{LRP8} \code{LRRC15} \code{LRRC17} \code{LRRC23} \code{LRRC32} \code{LRRC37A} \code{LRRC4C} \code{LRRC55} \code{LRRC59} \code{LRRIQ1} \code{LRRN1} \code{LRRN2} \code{LRRN3} \code{LRRTM2} \code{LRRTM4} \code{LSAMP} \code{LSM14B} \code{LST1} \code{LTBP2} \code{LTK} \code{LUM} \code{LY6H} \code{LY6K} \code{LY96} \code{LYVE1} \code{MACF1} \code{MAFB} \code{MAGEC1} \code{MAGED4B} \code{MAGEH1} \code{MAGEL2} \code{MAL} \code{MALAT1} \code{MAN1A1} \code{MAOB} \code{MAP3K14} \code{MAP6D1} \code{MAPK10} \code{MAPK12} \code{MAPK13} \code{MAPT} \code{MARCO} \code{MARK4} \code{MASP2} \code{MAT1A} \code{MATN3} \code{MATN4} \code{MB} \code{MBNL1} \code{MBOAT2} \code{MBP} \code{MC4R} \code{MCOLN3} \code{MCTP1} \code{MDFI} \code{MDFIC} \code{MDK} \code{ME1} \code{MEIS1} \code{MEIS2} \code{MELK} \code{MEP1B} \code{MEPE} \code{METTL7A} \code{MFAP2} \code{MFAP3L} \code{MFAP5} \code{MFSD11} \code{MFSD4} \code{MGAM} \code{MGAT4C} \code{MGLL} \code{MGMT} \code{MGP} \code{MGST1} \code{MIA2} \code{MLF1IP} \code{MLL} \code{MLLT4} \code{MLN} \code{MLPH} \code{MLXIPL} \code{MME} \code{MMP1} \code{MMP12} \code{MMP25} \code{MMP26} \code{MMP3} \code{MMP7} \code{MNDA} \code{MNT} \code{MNX1} \code{MOBKL1A} \code{MOBKL2B} \code{MOSC1} \code{MOXD1} \code{MPEG1} \code{MPL} \code{MPP6} \code{MPPED1} \code{MRC1} \code{MS4A4A} \code{MS4A6A} \code{MS4A7} \code{MSMB} \code{MSRB2} \code{MST4} \code{MSX1} \code{MT1H} \code{MT1L} \code{MT3} \code{MTMR9} \code{MTP18} \code{MTTP} \code{MTUS1} \code{MUC13} \code{MUC15} \code{MUC5AC} \code{MUSK} \code{MX2} \code{MXRA5} \code{MYLK} \code{MYO10} \code{MYO15B} \code{MYO1A} \code{MYO3A} \code{MYO5B} \code{MYOM1} \code{MYOM2} \code{MYRIP} \code{N4BP2L2} \code{N6AMT1} \code{NAA20} \code{NAAA} \code{NACAD} \code{NANOS1} \code{NAP1L3} \code{NAT8L} \code{NAV2} \code{NAV3} \code{NBEA} \code{NBLA00301} \code{NCRNA00081} \code{NCRNA00171} \code{NDC80} \code{NDN} \code{NDRG1} \code{NDRG4} \code{NDUFA4L2} \code{NDUFAF4} \code{NEAT1} \code{NECAB2} \code{NEFH} \code{NEFL} \code{NEFM} \code{NEK11} \code{NEK2} \code{NEK6} \code{NELL2} \code{NETO2} \code{NEURL1B} \code{NEURL3} \code{NEUROD1} \code{NEXN} \code{NFATC1} \code{NFATC2} \code{NFE2L2} \code{NHLH2} \code{NHSL1} \code{NIPAL2} \code{NISCH} \code{NKAPL} \code{NKD1} \code{NKX2-2} \code{NKX3-1} \code{NKX3-2} \code{NLGN1} \code{NME4} \code{NMNAT2} \code{NMU} \code{NMUR2} \code{NNMT} \code{NOL4} \code{NOV} \code{NOX5} \code{NPAS2} \code{NPAS3} \code{NPC1L1} \code{NPFF} \code{NPPB} \code{NPR3} \code{NPTN} \code{NPTX2} \code{NPW} \code{NPY} \code{NQO1} \code{NR0B1} \code{NR2F1} \code{NR3C1} \code{NR4A2} \code{NR5A2} \code{NRCAM} \code{NRGN} \code{NRN1} \code{NRP1} \code{NRXN1} \code{NRXN2} \code{NRXN3} \code{NT5DC3} \code{NT5E} \code{NTM} \code{NTN4} \code{NTRK2} \code{NTS} \code{NTSR1} \code{NUDT18} \code{NUDT3} \code{NUDT5} \code{NUP160} \code{NUP62CL} \code{NUPR1} \code{NXF2} \code{NXPH3} \code{NXPH4} \code{ODZ2} \code{OGDHL} \code{OGG1} \code{OIP5} \code{OIT3} \code{OLFM1} \code{OLFML2A} \code{OLIG1} \code{OLR1} \code{OMG} \code{ONECUT2} \code{OPA1} \code{OPCML} \code{OR1C1} \code{OR51E1} \code{OS9} \code{P2RX1} \code{P2RX2} \code{P2RY14} \code{P4HA2} \code{PABPC5} \code{PADI3} \code{PAFAH1B3} \code{PAH} \code{PAIP2B} \code{PAK3} \code{PAK6} \code{PAM} \code{PAPD4} \code{PAPSS1} \code{PAPSS2} \code{PAQR6} \code{PAR1} \code{PARM1} \code{PARP2} \code{PART1} \code{PARVB} \code{PATZ1} \code{PAX6} \code{PAX8} \code{PBK} \code{PBLD} \code{PBRM1} \code{PBXIP1} \code{PCDH11X} \code{PCDH12} \code{PCDH17} \code{PCDH18} \code{PCDH19} \code{PCDH20} \code{PCDH8} \code{PCDH9} \code{PCDHB16} \code{PCDHB17} \code{PCK1} \code{PCLO} \code{PCM1} \code{PCOLCE2} \code{PCP4} \code{PCSK1} \code{PCSK1N} \code{PCSK2} \code{PCSK5} \code{PDCD6} \code{PDE2A} \code{PDE3A} \code{PDE4DIP} \code{PDE6B} \code{PDE8B} \code{PDE9A} \code{PDGFC} \code{PDGFRL} \code{PDIA2} \code{PDK3} \code{PDK4} \code{PDLIM3} \code{PDLIM4} \code{PDPN} \code{PDS5B} \code{PDZD2} \code{PDZK1} \code{PDZK1IP1} \code{PDZRN3} \code{PDZRN4} \code{PEG10} \code{PELI1} \code{PEMT} \code{PENK} \code{PFKFB2} \code{PFKFB3} \code{PFN1} \code{PGAP1} \code{PGBD5} \code{PHACTR1} \code{PHGDH} \code{PHLDA1} \code{PHLDA2} \code{PHOX2A} \code{PI3} \code{PIF1} \code{PIGR} \code{PIK3R4} \code{PIP5K1B} \code{PIPOX} \code{PITX1} \code{PIWIL4} \code{PKNOX2} \code{PKP2} \code{PKP3} \code{PLA1A} \code{PLA2G1B} \code{PLA2G4A} \code{PLA2G5} \code{PLAG1} \code{PLAGL1} \code{PLBD1} \code{PLCB1} \code{PLCE1} \code{PLCL2} \code{PLCXD2} \code{PLEK2} \code{PLEKHB1} \code{PLEKHO2} \code{PLIN2} \code{PLIN3} \code{PLK2} \code{PLLP} \code{PLOD2} \code{PLUNC} \code{PLXDC1} \code{PLXNA2} \code{PLXNC1} \code{PMEPA1} \code{PMM1} \code{PNLIP} \code{PNLIPRP1} \code{PNLIPRP2} \code{PNMAL1} \code{PNMT} \code{POF1B} \code{POLA1} \code{POLR2G} \code{POMC} \code{PON1} \code{PON3} \code{POPDC3} \code{POSTN} \code{POU2AF1} \code{POU3F2} \code{PPARGC1B} \code{PPEF1} \code{PPM1E} \code{PPP1R14D} \code{PPP1R1A} \code{PPP2R2C} \code{PPP4R4} \code{PPY} \code{PPY2} \code{PRDM10} \code{PRELP} \code{PREX1} \code{PRKACB} \code{PRKAG2} \code{PRKCD} \code{PRKCE} \code{PRKG2} \code{PRKXP1} \code{PRLH} \code{PRLR} \code{PRMT8} \code{PRND} \code{PRODH2} \code{PROM1} \code{ProSAPiP1} \code{PRPH} \code{PRPSAP1} \code{PRR4} \code{PRRX2} \code{PRSS1} \code{PRSS2} \code{PRSS22} \code{PRSS3} \code{PRSS8} \code{PSAT1} \code{PSCA} \code{PSMA1} \code{PSMB9} \code{PSPC1} \code{PTCH1} \code{PTGDS} \code{PTGER3} \code{PTGFR} \code{PTGS1} \code{PTGS2} \code{PTHLH} \code{PTP4A1} \code{PTPLA} \code{PTPN20B} \code{PTPN7} \code{PTPRB} \code{PTPRC} \code{PTPRH} \code{PTPRJ} \code{PTPRK} \code{PTPRN2} \code{PTPRT} \code{PTPRZ1} \code{PTTG1} \code{PUS7} \code{PVR} \code{PVRL2} \code{PVRL3} \code{PVRL4} \code{PXDN} \code{PYCARD} \code{PYGL} \code{QDPR} \code{QPCT} \code{QPRT} \code{RAB11FIP5} \code{RAB17} \code{RAB20} \code{RAB26} \code{RAB27A} \code{RAB27B} \code{RAB33A} \code{RAB38} \code{RAB3B} \code{RAB3IL1} \code{RAB7L1} \code{RAB9B} \code{RABL5} \code{RAD51AP1} \code{RADIL} \code{RALYL} \code{RAMP1} \code{RAP1GAP2} \code{RAPGEF4} \code{RARB} \code{RARRES1} \code{RARRES2} \code{RARRES3} \code{RASAL1} \code{RASD1} \code{RASGRF1} \code{RASGRP1} \code{RBM16} \code{RBP1} \code{RBP4} \code{RBPJL} \code{RCAN2} \code{RDH16} \code{RDH5} \code{REG1B} \code{REG3A} \code{RELN} \code{REN} \code{RERGL} \code{RFPL1} \code{RGN} \code{RGS1} \code{RGS11} \code{RGS16} \code{RGS17} \code{RGS2} \code{RGS22} \code{RGS4} \code{RGS5} \code{RGS7} \code{RGS9} \code{RHD} \code{RHOBTB3} \code{RHOU} \code{RHPN2} \code{RIBC2} \code{RIN2} \code{RIN3} \code{RIPK2} \code{RLN2} \code{RNASE4} \code{RND1} \code{RND3} \code{RNF128} \code{RNF150} \code{RNF152} \code{RNF185} \code{RNF186} \code{RNF213} \code{RNLS} \code{ROBO1} \code{ROBO2} \code{ROBO3} \code{ROR2} \code{RORC} \code{RPH3AL} \code{RPL10A} \code{RPL4} \code{RPRM} \code{RPS2} \code{RRAS2} \code{RRBP1} \code{RRP1B} \code{RSAD2} \code{RTN1} \code{RTN2} \code{RTN4} \code{RUFY2} \code{RUFY3} \code{RUNX1T1} \code{RUNX2} \code{S100A14} \code{S100A4} \code{S100A8} \code{S100A9} \code{S100P} \code{S1PR4} \code{SAA1} \code{SAA4} \code{SCARB1} \code{SCD5} \code{SCG2} \code{SCG3} \code{SCG5} \code{SCGB2A1} \code{SCGB3A1} \code{SCGN} \code{SCN1B} \code{SCN3B} \code{SCRG1} \code{SCTR} \code{SDC2} \code{SDC4} \code{SDCBP2} \code{SDCCAG8} \code{SDF2} \code{SDHAF2} \code{SEC24D} \code{SECTM1} \code{SELE} \code{SELENBP1} \code{SELP} \code{SEMA3E} \code{SEMA3G} \code{SEMA5A} \code{SEMA6A} \code{SEMA6D} \code{SERPINA1} \code{SERPINA10} \code{SERPINA3} \code{SERPINA5} \code{SERPINA6} \code{SERPINA7} \code{SERPINB2} \code{SERPINB9} \code{SERPINC1} \code{SERPIND1} \code{SERPINE1} \code{SERPINF1} \code{SERPINI1} \code{SERPINI2} \code{SERTAD4} \code{SETD5} \code{SEZ6L} \code{SFN} \code{SFRP1} \code{SFRP2} \code{SFRP4} \code{SFRS2} \code{SGCD} \code{SGCG} \code{SGEF} \code{SGK1} \code{SGK223} \code{SH2D2A} \code{SH3BP4} \code{SH3GL2} \code{SH3GL3} \code{SH3PXD2A} \code{SHOX2} \code{SHROOM2} \code{SI} \code{SIGLEC9} \code{SIRT6} \code{SKAP1} \code{SKAP2} \code{SLA} \code{SLA2} \code{SLC12A5} \code{SLC12A7} \code{SLC12A8} \code{SLC14A1} \code{SLC16A3} \code{SLC16A7} \code{SLC16A9} \code{SLC17A1} \code{SLC17A6} \code{SLC18A1} \code{SLC18A2} \code{SLC18A3} \code{SLC19A1} \code{SLC1A2} \code{SLC1A5} \code{SLC22A17} \code{SLC22A23} \code{SLC22A5} \code{SLC23A2} \code{SLC25A27} \code{SLC25A34} \code{SLC26A4} \code{SLC27A2} \code{SLC2A1} \code{SLC2A2} \code{SLC30A2} \code{SLC35F2} \code{SLC37A4} \code{SLC38A1} \code{SLC38A4} \code{SLC39A8} \code{SLC3A1} \code{SLC3A2} \code{SLC40A1} \code{SLC43A1} \code{SLC5A1} \code{SLC6A14} \code{SLC6A17} \code{SLC6A19} \code{SLC6A3} \code{SLC6A4} \code{SLC6A7} \code{SLC6A8} \code{SLC7A2} \code{SLC7A7} \code{SLC7A8} \code{SLC7A9} \code{SLC8A1} \code{SLC8A2} \code{SLC9A3} \code{SLCO1A2} \code{SLCO1B1} \code{SLCO2A1} \code{SLCO3A1} \code{SLCO5A1} \code{SLIT1} \code{SLIT2} \code{SLITRK5} \code{SLITRK6} \code{SMAD3} \code{SMAD4} \code{SMAGP} \code{SMARCA1} \code{SMARCD3} \code{SMC6} \code{SMEK1} \code{SMO} \code{SMOC1} \code{SMOC2} \code{SMPD3} \code{SMPDL3A} \code{SMURF1} \code{SMYD2} \code{SMYD3} \code{SNAP91} \code{SNCA} \code{SNED1} \code{SNRPD3} \code{SNRPN} \code{SNTB1} \code{SOBP} \code{SOD2} \code{SOD3} \code{SORBS2} \code{SORCS2} \code{SORD} \code{SOSTDC1} \code{SOX1} \code{SOX11} \code{SOX15} \code{SOX4} \code{SOX9} \code{SPAG17} \code{SPAG4} \code{SPAG6} \code{SPARCL1} \code{SPATS2L} \code{SPC25} \code{SPEF1} \code{SPG11} \code{SPIN1} \code{SPINK2} \code{SPINK4} \code{SPINK5} \code{SPOCK1} \code{SPOCK2} \code{SPOCK3} \code{SPON1} \code{SPP1} \code{SPRR1A} \code{SPSB3} \code{SPTBN1} \code{SRD5A3} \code{SRGAP3} \code{SRM} \code{SRPX} \code{SRPX2} \code{SSBP1} \code{SSBP2} \code{SSC5D} \code{SST} \code{SSTR1} \code{SSTR2} \code{SSX2IP} \code{ST18} \code{ST3GAL1} \code{ST3GAL6} \code{ST6GAL1} \code{ST6GALNAC5} \code{ST8SIA4} \code{STAC2} \code{STAMBPL1} \code{STAP2} \code{STAR} \code{STAT1} \code{STAT4} \code{STAU2} \code{STC1} \code{STC2} \code{STEAP1} \code{STEAP2} \code{STEAP3} \code{STK32B} \code{STK38L} \code{STMN1} \code{STMN2} \code{STOX2} \code{STXBP6} \code{STYXL1} \code{SULF2} \code{SULT1E1} \code{SULT2A1} \code{SUSD5} \code{SV2A} \code{SV2B} \code{SV2C} \code{SVEP1} \code{SVIL} \code{SVOP} \code{SYNGR4} \code{SYNJ1} \code{SYNJ2} \code{SYNM} \code{SYNPO} \code{SYT1} \code{SYT13} \code{SYT17} \code{SYT4} \code{SYT5} \code{SYT7} \code{SYT9} \code{TAC1} \code{TACSTD2} \code{TAGLN3} \code{TANC1} \code{TAP2} \code{TAPBPL} \code{TARDBP} \code{TAT} \code{TBC1D16} \code{TBC1D24} \code{TBPL1} \code{TBX19} \code{TCEA3} \code{TCEAL7} \code{TCEB3} \code{TCF7} \code{TCN1} \code{TCN2} \code{TDO2} \code{TDRD12} \code{TECR} \code{TEKT2} \code{TERT} \code{TET2} \code{TEX15} \code{TFCP2L1} \code{TFF1} \code{TFF2} \code{TFF3} \code{TFPI2} \code{TGFA} \code{TGFB2} \code{TGFB3} \code{TGFBR3} \code{TGIF1} \code{TH} \code{TH1L} \code{THBS1} \code{THBS2} \code{THBS4} \code{THUMPD1} \code{THY1} \code{TIA1} \code{TIAM1} \code{TIMM13} \code{TIMP1} \code{TIMP2} \code{TJP2} \code{TKTL1} \code{TLE2} \code{TLL2} \code{TLR2} \code{TLR3} \code{TM4SF1} \code{TM4SF4} \code{TM4SF5} \code{TMC5} \code{TMCO7} \code{TMED8} \code{TMEFF2} \code{TMEM100} \code{TMEM132A} \code{TMEM150C} \code{TMEM158} \code{TMEM165} \code{TMEM176A} \code{TMEM176B} \code{TMEM181} \code{TMEM204} \code{TMEM27} \code{TMEM30B} \code{TMEM35} \code{TMEM37} \code{TMEM45A} \code{TMEM47} \code{TMEM51} \code{TMEM53} \code{TMEM90B} \code{TMEM98} \code{TMOD1} \code{TMPRSS11E} \code{TMPRSS15} \code{TMPRSS2} \code{TMPRSS3} \code{TMPRSS4} \code{TMSB15A} \code{TMTC1} \code{TNFAIP2} \code{TNFAIP3} \code{TNFAIP6} \code{TNFAIP8L1} \code{TNFRSF11B} \code{TNFRSF12A} \code{TNFRSF14} \code{TNFRSF21} \code{TNFSF10} \code{TNFSF8} \code{TNS3} \code{TOP2A} \code{TOR1AIP1} \code{TOX3} \code{TP53I11} \code{TP53TG3} \code{TPCN1} \code{TPD52} \code{TPH1} \code{TPK1} \code{TPM2} \code{TPPP3} \code{TPST2} \code{TPTE} \code{TREM2} \code{TRH} \code{TRHR} \code{TRIB3} \code{TRIM2} \code{TRIM9} \code{TRIP13} \code{TRPC4} \code{TRPM5} \code{TRPV5} \code{TSC2} \code{TSHR} \code{TSHZ3} \code{TSPAN1} \code{TSPAN12} \code{TSPAN13} \code{TSPAN2} \code{TSPAN5} \code{TSPAN7} \code{TSPAN8} \code{TSPO} \code{TSPYL1} \code{TSPYL2} \code{TSPYL4} \code{TSPYL5} \code{TST} \code{TSTD2} \code{TTC39B} \code{TTPA} \code{TTR} \code{TTTY15} \code{TUB} \code{TUBB1} \code{TWIST1} \code{TYMS} \code{UACA} \code{UBE2C} \code{UCHL1} \code{UCP2} \code{UGT1A6} \code{UGT2A3} \code{UGT2B17} \code{UHRF1} \code{USH1C} \code{USP18} \code{USP29} \code{USP34} \code{UST} \code{VAMP8} \code{VAT1L} \code{VAV3} \code{VCAM1} \code{VCX3A} \code{VDR} \code{VEGFC} \code{VGF} \code{VGLL3} \code{VIL1} \code{VIP} \code{VIPR1} \code{VIPR2} \code{VLDLR} \code{VNN1} \code{VNN2} \code{VPS37B} \code{VPS54} \code{VSNL1} \code{VTN} \code{WAPAL} \code{WASL} \code{WDR47} \code{WFDC1} \code{WFDC2} \code{WFS1} \code{WISP2} \code{WLS} \code{WNK2} \code{WNT10A} \code{WNT11} \code{WNT4} \code{WNT5A} \code{WNT7A} \code{WSCD1} \code{XAGE1D} \code{XIST} \code{YPEL2} \code{ZBBX} \code{ZBTB16} \code{ZBTB20} \code{ZBTB26} \code{ZCCHC8} \code{ZDBF2} \code{ZFHX3} \code{ZFR2} \code{ZMYND10} \code{ZNF189} \code{ZNF219} \code{ZNF239} \code{ZNF300} \code{ZNF331} \code{ZNF394} \code{ZNF446} \code{ZNF521} \code{ZNF671} \code{ZNF710} \code{ZNF784} \code{ZNF880} \code{ZNRD1} \code{ZRANB3} \code{ZSCAN18}}
    \item{\code{X131}}{a numeric vector}
    \item{\code{X134}}{a numeric vector}
    \item{\code{X136}}{a numeric vector}
    \item{\code{X137}}{a numeric vector}
    \item{\code{X132}}{a numeric vector}
    \item{\code{X135}}{a numeric vector}
    \item{\code{X138}}{a numeric vector}
    \item{\code{X90}}{a numeric vector}
    \item{\code{X102}}{a numeric vector}
    \item{\code{X112}}{a numeric vector}
    \item{\code{X119}}{a numeric vector}
    \item{\code{X55}}{a numeric vector}
    \item{\code{X36}}{a numeric vector}
    \item{\code{X38}}{a numeric vector}
    \item{\code{X39}}{a numeric vector}
    \item{\code{X40}}{a numeric vector}
    \item{\code{X44}}{a numeric vector}
    \item{\code{X49}}{a numeric vector}
    \item{\code{X51}}{a numeric vector}
    \item{\code{X52}}{a numeric vector}
    \item{\code{X56}}{a numeric vector}
    \item{\code{X57}}{a numeric vector}
    \item{\code{X58}}{a numeric vector}
    \item{\code{X62}}{a numeric vector}
    \item{\code{X68}}{a numeric vector}
    \item{\code{X69}}{a numeric vector}
    \item{\code{X72}}{a numeric vector}
    \item{\code{X73}}{a numeric vector}
    \item{\code{X74}}{a numeric vector}
    \item{\code{X75}}{a numeric vector}
    \item{\code{X84}}{a numeric vector}
    \item{\code{X93}}{a numeric vector}
    \item{\code{X98}}{a numeric vector}
    \item{\code{X114}}{a numeric vector}
    \item{\code{X116}}{a numeric vector}
    \item{\code{X120}}{a numeric vector}
    \item{\code{X122}}{a numeric vector}
    \item{\code{X124}}{a numeric vector}
    \item{\code{X126}}{a numeric vector}
    \item{\code{X130}}{a numeric vector}
    \item{\code{X133}}{a numeric vector}
    \item{\code{X35}}{a numeric vector}
    \item{\code{X41}}{a numeric vector}
    \item{\code{X42}}{a numeric vector}
    \item{\code{X59}}{a numeric vector}
    \item{\code{X61}}{a numeric vector}
    \item{\code{X121}}{a numeric vector}
    \item{\code{X123}}{a numeric vector}
    \item{\code{X125}}{a numeric vector}
    \item{\code{X127}}{a numeric vector}
    \item{\code{X33}}{a numeric vector}
    \item{\code{X64}}{a numeric vector}
    \item{\code{X70}}{a numeric vector}
    \item{\code{X87}}{a numeric vector}
    \item{\code{X99}}{a numeric vector}
    \item{\code{X115}}{a numeric vector}
    \item{\code{X129}}{a numeric vector}
    \item{\code{X47}}{a numeric vector}
    \item{\code{X71}}{a numeric vector}
    \item{\code{X92}}{a numeric vector}
    \item{\code{X94}}{a numeric vector}
    \item{\code{X100}}{a numeric vector}
    \item{\code{X118}}{a numeric vector}
    \item{\code{X34}}{a numeric vector}
    \item{\code{X37}}{a numeric vector}
    \item{\code{X46}}{a numeric vector}
    \item{\code{X48}}{a numeric vector}
    \item{\code{X50}}{a numeric vector}
    \item{\code{X53}}{a numeric vector}
    \item{\code{X54}}{a numeric vector}
    \item{\code{X60}}{a numeric vector}
    \item{\code{X63}}{a numeric vector}
    \item{\code{X65}}{a numeric vector}
    \item{\code{X66}}{a numeric vector}
    \item{\code{X67}}{a numeric vector}
    \item{\code{X86}}{a numeric vector}
    \item{\code{X89}}{a numeric vector}
    \item{\code{X95}}{a numeric vector}
    \item{\code{X96}}{a numeric vector}
    \item{\code{X97}}{a numeric vector}
    \item{\code{X117}}{a numeric vector}
    \item{\code{X128}}{a numeric vector}
    \item{\code{X101}}{a numeric vector}
    \item{\code{X108}}{a numeric vector}
    \item{\code{X109}}{a numeric vector}
    \item{\code{X110}}{a numeric vector}
  }
}
\details{
MLP vs other dataset
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(MLPvsother.Indata)
## maybe str(MLPvsother.Indata) ; plot(MLPvsother.Indata) ...
}
\keyword{datasets}
