\name{MLPvsother.Indlabdata}
\alias{MLPvsother.Indlabdata}
\docType{data}
\title{
Mlp vs others NMF labels.
}
\description{
Mlp vs others NMF labels.
}
\usage{data("MLPvsother.Indlabdata")}
\format{
  A data frame with 86 observations on the following 2 variables.
  \describe{
    \item{\code{Name}}{a character vector}
    \item{\code{membership.ordered}}{a numeric vector}
  }
}
\details{
Mlp vs others NMF labels
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(MLPvsother.Indlabdata)
## maybe str(MLPvsother.Indlabdata) ; plot(MLPvsother.Indlabdata) ...
}
\keyword{datasets}
