

\name{PanNETassigner-package}
\alias{PanNETassigner-package}
\alias{PanNETassigner}
\docType{package}
\title{
\packageTitle{PanNETassigner}
}
\description{
\packageDescription{PanNETassigner}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PanNETassigner}
\packageIndices{PanNETassigner}
This is a package to reproduce results in Sadanandam, et all (2015), Cancer Discovery.
}

\author{
\packageAuthor{PanNETassigner}

Maintainer: \packageMaintainer{PanNETassigner}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}

