\name{PanNETsubtypesmicRNA.Indata}
\alias{PanNETsubtypesmicRNA.Indata}
\docType{data}
\title{micRNA expression data
}
\description{
micRNA expression data of tumors and islets used to identify PanNET subtypes. 
}
\usage{data("PanNETsubtypesmicRNA.Indata")}
\format{
  A data frame with 207 observations on the following 41 variables.
  \describe{
    \item{\code{mir...1.}}{a factor with levels \code{#N/A} \code{MIR1-1} \code{MIR1-2} \code{MIR100} \code{MIR101-1} \code{MIR101-2} \code{MIR103-1} \code{MIR103-2} \code{MIR105-1} \code{MIR105-2} \code{MIR106A} \code{MIR106B} \code{MIR107} \code{MIR10A} \code{MIR10B} \code{MIR122} \code{MIR124-1} \code{MIR124-2} \code{MIR124-3} \code{MIR125A} \code{MIR125B1} \code{MIR125B2} \code{MIR126} \code{MIR127} \code{MIR128-1} \code{MIR128-2} \code{MIR129-1} \code{MIR129-2} \code{MIR130A} \code{MIR130B} \code{MIR132} \code{MIR133A1} \code{MIR133A2} \code{MIR133B} \code{MIR134} \code{MIR135A1} \code{MIR135A2} \code{MIR135B} \code{MIR136} \code{MIR137} \code{MIR138-1} \code{MIR138-2} \code{MIR139} \code{MIR140} \code{MIR141} \code{MIR142} \code{MIR143} \code{MIR144} \code{MIR145} \code{MIR146A} \code{MIR147} \code{MIR148A} \code{MIR148B} \code{MIR149} \code{MIR150} \code{MIR151} \code{MIR152} \code{MIR153-1} \code{MIR153-2} \code{MIR154} \code{MIR155} \code{MIR15A} \code{MIR15B} \code{MIR16-1} \code{MIR16-2} \code{MIR17} \code{MIR181A1} \code{MIR181A2} \code{MIR181B1} \code{MIR181B2} \code{MIR181C} \code{MIR182} \code{MIR183} \code{MIR184} \code{MIR185} \code{MIR186} \code{MIR187} \code{MIR188} \code{MIR18A} \code{MIR190} \code{MIR191} \code{MIR192} \code{MIR193A} \code{MIR194-1} \code{MIR194-2} \code{MIR195} \code{MIR196A1} \code{MIR196A2} \code{MIR196B} \code{MIR197} \code{MIR198} \code{MIR199A1} \code{MIR199A2} \code{MIR199B} \code{MIR19A} \code{MIR19B1} \code{MIR19B2} \code{MIR200A} \code{MIR200B} \code{MIR200C} \code{MIR203} \code{MIR204} \code{MIR205} \code{MIR206} \code{MIR208A} \code{MIR20A} \code{MIR21} \code{MIR210} \code{MIR211} \code{MIR212} \code{MIR214} \code{MIR215} \code{MIR216A} \code{MIR217} \code{MIR218-1} \code{MIR218-2} \code{MIR219-1} \code{MIR219-2} \code{MIR22} \code{MIR220A} \code{MIR221} \code{MIR222} \code{MIR223} \code{MIR224} \code{MIR23A} \code{MIR23B} \code{MIR24-1} \code{MIR24-2} \code{MIR25} \code{MIR26A1} \code{MIR26A2} \code{MIR26B} \code{MIR27A} \code{MIR27B} \code{MIR28} \code{MIR296} \code{MIR299} \code{MIR29A} \code{MIR29B1} \code{MIR29B2} \code{MIR29C} \code{MIR301A} \code{MIR302A} \code{MIR302B} \code{MIR302C} \code{MIR302D} \code{MIR30A} \code{MIR30B} \code{MIR30C1} \code{MIR30C2} \code{MIR30D} \code{MIR30E} \code{MIR31} \code{MIR32} \code{MIR320A} \code{MIR323} \code{MIR324} \code{MIR325} \code{MIR326} \code{MIR328} \code{MIR330} \code{MIR331} \code{MIR335} \code{MIR337} \code{MIR338} \code{MIR339} \code{MIR33A} \code{MIR340} \code{MIR342} \code{MIR345} \code{MIR346} \code{MIR34A} \code{MIR34B} \code{MIR34C} \code{MIR367} \code{MIR369} \code{MIR370} \code{MIR371} \code{MIR372} \code{MIR373} \code{MIR374A} \code{MIR376C} \code{MIR7-1} \code{MIR7-2} \code{MIR7-3} \code{MIR9-1} \code{MIR9-2} \code{MIR9-3} \code{MIR92A1} \code{MIR92A2} \code{MIR93} \code{MIR95} \code{MIR96} \code{MIR98} \code{MIR99A} \code{MIR99B} \code{MIRLET7A1} \code{MIRLET7A2} \code{MIRLET7A3} \code{MIRLET7B} \code{MIRLET7C} \code{MIRLET7D} \code{MIRLET7E} \code{MIRLET7F1} \code{MIRLET7F2} \code{MIRLET7G} \code{MIRLET7I}}
    \item{\code{X220}}{a numeric vector}
    \item{\code{X221}}{a numeric vector}
    \item{\code{X222}}{a numeric vector}
    \item{\code{X223}}{a numeric vector}
    \item{\code{X224}}{a numeric vector}
    \item{\code{X225}}{a numeric vector}
    \item{\code{X226}}{a numeric vector}
    \item{\code{X227}}{a numeric vector}
    \item{\code{X228}}{a numeric vector}
    \item{\code{X229}}{a numeric vector}
    \item{\code{X230}}{a numeric vector}
    \item{\code{X231}}{a numeric vector}
    \item{\code{X232}}{a numeric vector}
    \item{\code{X233}}{a numeric vector}
    \item{\code{X234}}{a numeric vector}
    \item{\code{X235}}{a numeric vector}
    \item{\code{X236}}{a numeric vector}
    \item{\code{X237}}{a numeric vector}
    \item{\code{X238}}{a numeric vector}
    \item{\code{X239}}{a numeric vector}
    \item{\code{X240}}{a numeric vector}
    \item{\code{X241}}{a numeric vector}
    \item{\code{X242}}{a numeric vector}
    \item{\code{X247}}{a numeric vector}
    \item{\code{X249}}{a numeric vector}
    \item{\code{X250}}{a numeric vector}
    \item{\code{X251}}{a numeric vector}
    \item{\code{X252}}{a numeric vector}
    \item{\code{X253}}{a numeric vector}
    \item{\code{X254}}{a numeric vector}
    \item{\code{X255}}{a numeric vector}
    \item{\code{X256}}{a numeric vector}
    \item{\code{X265}}{a numeric vector}
    \item{\code{X273}}{a numeric vector}
    \item{\code{X274}}{a numeric vector}
    \item{\code{X275}}{a numeric vector}
    \item{\code{X276}}{a numeric vector}
    \item{\code{X277}}{a numeric vector}
    \item{\code{X278}}{a numeric vector}
    \item{\code{X279}}{a numeric vector}
  }
}
\details{
This is the micRNA expression dataset of tumors and islets used to identify PanNET subtypes. micRNAs are on the rows and the first column contain the micRNAs names.
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(PanNETsubtypesmicRNA.Indata)
## maybe str(PanNETsubtypesmicRNA.Indata) ; plot(PanNETsubtypesmicRNA.Indata) ...
}
\keyword{datasets}
