\name{PanNETsubtypesmicRNAVal.Indata}
\alias{PanNETsubtypesmicRNAVal.Indata}
\docType{data}
\title{
micRNA expression validation data
}
\description{
This a micRNA expression dataset for validation of PanNET subtypes.
}
\usage{data("PanNETsubtypesmicRNAVal.Indata")}
\format{
  A data frame with 1368 observations on the following 51 variables.
  \describe{
    \item{\code{mIRs}}{a factor with levels \code{bkv-miR-B1-3p} \code{bkv-miR-B1-5p} \code{Blank} \code{dmr_285} \code{dmr_3} \code{dmr_308} \code{dmr_316} \code{dmr_31a} \code{dmr_6} \code{ebv-miR-BART1-3p} \code{ebv-miR-BART1-5p} \code{ebv-miR-BART10} \code{ebv-miR-BART10*} \code{ebv-miR-BART11-3p} \code{ebv-miR-BART11-5p} \code{ebv-miR-BART12} \code{ebv-miR-BART13} \code{ebv-miR-BART13*} \code{ebv-miR-BART14} \code{ebv-miR-BART14*} \code{ebv-miR-BART15} \code{ebv-miR-BART16} \code{ebv-miR-BART17-3p} \code{ebv-miR-BART17-5p} \code{ebv-miR-BART18-3p} \code{ebv-miR-BART18-5p} \code{ebv-miR-BART19-3p} \code{ebv-miR-BART19-5p} \code{ebv-miR-BART2-3p} \code{ebv-miR-BART2-5p} \code{ebv-miR-BART20-3p} \code{ebv-miR-BART20-5p} \code{ebv-miR-BART21-3p} \code{ebv-miR-BART21-5p} \code{ebv-miR-BART22} \code{ebv-miR-BART3} \code{ebv-miR-BART3*} \code{ebv-miR-BART4} \code{ebv-miR-BART4*} \code{ebv-miR-BART5} \code{ebv-miR-BART5*} \code{ebv-miR-BART6-3p} \code{ebv-miR-BART6-5p} \code{ebv-miR-BART7} \code{ebv-miR-BART7*} \code{ebv-miR-BART8} \code{ebv-miR-BART8*} \code{ebv-miR-BART9} \code{ebv-miR-BART9*} \code{ebv-miR-BHRF1-1} \code{ebv-miR-BHRF1-2} \code{ebv-miR-BHRF1-2*} \code{ebv-miR-BHRF1-3} \code{hbv-miR-B20} \code{hbv-miR-B2RC} \code{hbv-miR-B4} \code{hcmv-miR-UL112} \code{hcmv-miR-UL148D} \code{hcmv-miR-UL22A} \code{hcmv-miR-UL22A*} \code{hcmv-miR-UL36} \code{hcmv-miR-UL36*} \code{hcmv-miR-UL70-3p} \code{hcmv-miR-UL70-5p} \code{hcmv-miR-US25-1} \code{hcmv-miR-US25-1*} \code{hcmv-miR-US25-2-3p} \code{hcmv-miR-US25-2-5p} \code{hcmv-miR-US33-3p} \code{hcmv-miR-US33-5p} \code{hcmv-miR-US4} \code{hcmv-miR-US5-1} \code{hcmv-miR-US5-2} \code{hiv1-miR-H1} \code{hiv1-miR-N367} \code{hiv1-miR-TAR-3p} \code{hiv1-miR-TAR-5p} \code{hsa-let-7a-3p} \code{hsa-let-7a-5p} \code{hsa-let-7b-3p} \code{hsa-let-7b-5p} \code{hsa-let-7c} \code{hsa-let-7c*_v16.0} \code{hsa-let-7d-3p} \code{hsa-let-7d-5p} \code{hsa-let-7e-3p} \code{hsa-let-7e-5p} \code{hsa-let-7f-1-3p} \code{hsa-let-7f-2-3p} \code{hsa-let-7f-5p} \code{hsa-let-7g-3p} \code{hsa-let-7g-5p} \code{hsa-let-7i-3p} \code{hsa-let-7i-5p} \code{hsa-miR-1} \code{hsa-miR-100-3p} \code{hsa-miR-100-5p} \code{hsa-miR-101-3p} \code{hsa-miR-101-5p} \code{hsa-miR-103a-2-5p} \code{hsa-miR-103a-3p} \code{hsa-miR-103b} \code{hsa-miR-105-3p} \code{hsa-miR-105-5p} \code{hsa-miR-106a-3p} \code{hsa-miR-106b-3p} \code{hsa-miR-106b-5p} \code{hsa-miR-107} \code{hsa-miR-10a-3p} \code{hsa-miR-10a-5p} \code{hsa-miR-10b-3p} \code{hsa-miR-10b-5p} \code{hsa-miR-1178} \code{hsa-miR-1179} \code{hsa-miR-1180} \code{hsa-miR-1181} \code{hsa-miR-1182} \code{hsa-miR-1183} \code{hsa-miR-1184} \code{hsa-miR-1185-5p} \code{hsa-miR-1193} \code{hsa-miR-1197} \code{hsa-miR-1200} \code{hsa-miR-1202} \code{hsa-miR-1203} \code{hsa-miR-1204} \code{hsa-miR-1205} \code{hsa-miR-1206} \code{hsa-miR-1207-3p} \code{hsa-miR-1207-5p} \code{hsa-miR-1208} \code{hsa-miR-122-3p} \code{hsa-miR-122-5p} \code{hsa-miR-1224-3p} \code{hsa-miR-1224-5p} \code{hsa-miR-1225-3p} \code{hsa-miR-1225-5p} \code{hsa-miR-1226-3p} \code{hsa-miR-1226-5p} \code{hsa-miR-1227} \code{hsa-miR-1228-3p} \code{hsa-miR-1228-5p} \code{hsa-miR-1229} \code{hsa-miR-1231} \code{hsa-miR-1233} \code{hsa-miR-1234} \code{hsa-miR-1236} \code{hsa-miR-1237} \code{hsa-miR-1238} \code{hsa-miR-124-3p} \code{hsa-miR-124-5p} \code{hsa-miR-1243} \code{hsa-miR-1244} \code{hsa-miR-1245a} \code{hsa-miR-1246} \code{hsa-miR-1247-5p} \code{hsa-miR-1248} \code{hsa-miR-1249} \code{hsa-miR-1250} \code{hsa-miR-1251} \code{hsa-miR-1252} \code{hsa-miR-1253} \code{hsa-miR-1254} \code{hsa-miR-1255a} \code{hsa-miR-1255b-5p} \code{hsa-miR-1256} \code{hsa-miR-1257} \code{hsa-miR-1258} \code{hsa-miR-125a-3p} \code{hsa-miR-125a-5p} \code{hsa-miR-125b-1-3p} \code{hsa-miR-125b-2-3p} \code{hsa-miR-125b-5p} \code{hsa-miR-126-3p} \code{hsa-miR-126-5p} \code{hsa-miR-1260a} \code{hsa-miR-1260b} \code{hsa-miR-1261} \code{hsa-miR-1262} \code{hsa-miR-1263} \code{hsa-miR-1264} \code{hsa-miR-1265} \code{hsa-miR-1266} \code{hsa-miR-1267} \code{hsa-miR-1268a} \code{hsa-miR-1269a} \code{hsa-miR-127-3p} \code{hsa-miR-127-5p} \code{hsa-miR-1270} \code{hsa-miR-1271-5p} \code{hsa-miR-1272} \code{hsa-miR-1273a} \code{hsa-miR-1273c} \code{hsa-miR-1273d} \code{hsa-miR-1273e} \code{hsa-miR-1274a_v16.0} \code{hsa-miR-1274b_v16.0} \code{hsa-miR-1275} \code{hsa-miR-1276} \code{hsa-miR-1277-3p} \code{hsa-miR-1278} \code{hsa-miR-1279} \code{hsa-miR-128} \code{hsa-miR-1280} \code{hsa-miR-1281} \code{hsa-miR-1282} \code{hsa-miR-1283} \code{hsa-miR-1284} \code{hsa-miR-1285-3p} \code{hsa-miR-1286} \code{hsa-miR-1287} \code{hsa-miR-1288} \code{hsa-miR-1289} \code{hsa-miR-129-1-3p} \code{hsa-miR-129-2-3p} \code{hsa-miR-129-5p} \code{hsa-miR-1290} \code{hsa-miR-1291} \code{hsa-miR-1292} \code{hsa-miR-1293} \code{hsa-miR-1294} \code{hsa-miR-1295a} \code{hsa-miR-1296} \code{hsa-miR-1297} \code{hsa-miR-1298} \code{hsa-miR-1299} \code{hsa-miR-1301} \code{hsa-miR-1302} \code{hsa-miR-1303} \code{hsa-miR-1304-5p} \code{hsa-miR-1305} \code{hsa-miR-1306-3p} \code{hsa-miR-1307-3p} \code{hsa-miR-130a-3p} \code{hsa-miR-130a-5p} \code{hsa-miR-130b-3p} \code{hsa-miR-130b-5p} \code{hsa-miR-132-3p} \code{hsa-miR-132-5p} \code{hsa-miR-1321} \code{hsa-miR-1322} \code{hsa-miR-1323} \code{hsa-miR-1324} \code{hsa-miR-133a} \code{hsa-miR-133b} \code{hsa-miR-134} \code{hsa-miR-135a-3p} \code{hsa-miR-135a-5p} \code{hsa-miR-135b-3p} \code{hsa-miR-135b-5p} \code{hsa-miR-136-3p} \code{hsa-miR-136-5p} \code{hsa-miR-137} \code{hsa-miR-138-1-3p} \code{hsa-miR-138-2-3p} \code{hsa-miR-138-5p} \code{hsa-miR-139-3p} \code{hsa-miR-139-5p} \code{hsa-miR-140-3p} \code{hsa-miR-140-5p} \code{hsa-miR-141-3p} \code{hsa-miR-141-5p} \code{hsa-miR-142-3p} \code{hsa-miR-142-5p} \code{hsa-miR-143-3p} \code{hsa-miR-143-5p} \code{hsa-miR-144-3p} \code{hsa-miR-144-5p} \code{hsa-miR-145-3p} \code{hsa-miR-145-5p} \code{hsa-miR-1468} \code{hsa-miR-1469} \code{hsa-miR-146a-3p} \code{hsa-miR-146a-5p} \code{hsa-miR-146b-3p} \code{hsa-miR-146b-5p} \code{hsa-miR-1470} \code{hsa-miR-1471} \code{hsa-miR-147a} \code{hsa-miR-147b} \code{hsa-miR-148a-3p} \code{hsa-miR-148a-5p} \code{hsa-miR-148b-3p} \code{hsa-miR-148b-5p} \code{hsa-miR-149-3p} \code{hsa-miR-149-5p} \code{hsa-miR-150-3p} \code{hsa-miR-150-5p} \code{hsa-miR-151a-3p} \code{hsa-miR-151a-5p} \code{hsa-miR-152} \code{hsa-miR-153} \code{hsa-miR-1537} \code{hsa-miR-1538} \code{hsa-miR-1539} \code{hsa-miR-154-3p} \code{hsa-miR-154-5p} \code{hsa-miR-155-3p} \code{hsa-miR-155-5p} \code{hsa-miR-15a-3p} \code{hsa-miR-15a-5p} \code{hsa-miR-15b-3p} \code{hsa-miR-15b-5p} \code{hsa-miR-16-1-3p} \code{hsa-miR-16-2-3p} \code{hsa-miR-16-5p} \code{hsa-miR-17-3p} \code{hsa-miR-17-5p} \code{hsa-miR-181a-2-3p} \code{hsa-miR-181a-3p} \code{hsa-miR-181a-5p} \code{hsa-miR-181b-5p} \code{hsa-miR-181c-3p} \code{hsa-miR-181c-5p} \code{hsa-miR-181d} \code{hsa-miR-182-3p} \code{hsa-miR-182-5p} \code{hsa-miR-1825} \code{hsa-miR-1827} \code{hsa-miR-183-3p} \code{hsa-miR-183-5p} \code{hsa-miR-184} \code{hsa-miR-185-3p} \code{hsa-miR-185-5p} \code{hsa-miR-186-3p} \code{hsa-miR-186-5p} \code{hsa-miR-187-3p} \code{hsa-miR-187-5p} \code{hsa-miR-188-3p} \code{hsa-miR-188-5p} \code{hsa-miR-18a-3p} \code{hsa-miR-18a-5p} \code{hsa-miR-18b-3p} \code{hsa-miR-18b-5p} \code{hsa-miR-1908} \code{hsa-miR-1909-3p} \code{hsa-miR-1909-5p} \code{hsa-miR-190a} \code{hsa-miR-190b} \code{hsa-miR-191-3p} \code{hsa-miR-191-5p} \code{hsa-miR-1910} \code{hsa-miR-1911-3p} \code{hsa-miR-1911-5p} \code{hsa-miR-1912} \code{hsa-miR-1913} \code{hsa-miR-1914-3p} \code{hsa-miR-1914-5p} \code{hsa-miR-1915-3p} \code{hsa-miR-1915-5p} \code{hsa-miR-192-3p} \code{hsa-miR-192-5p} \code{hsa-miR-193a-3p} \code{hsa-miR-193a-5p} \code{hsa-miR-193b-3p} \code{hsa-miR-193b-5p} \code{hsa-miR-194-3p} \code{hsa-miR-194-5p} \code{hsa-miR-195-3p} \code{hsa-miR-195-5p} \code{hsa-miR-196a-3p} \code{hsa-miR-196a-5p} \code{hsa-miR-196b-3p} \code{hsa-miR-196b-5p} \code{hsa-miR-197-3p} \code{hsa-miR-1972} \code{hsa-miR-1973} \code{hsa-miR-1976} \code{hsa-miR-198} \code{hsa-miR-199a-3p} \code{hsa-miR-199a-5p} \code{hsa-miR-199b-5p} \code{hsa-miR-19a-3p} \code{hsa-miR-19a-5p} \code{hsa-miR-19b-1-5p} \code{hsa-miR-19b-2-5p} \code{hsa-miR-19b-3p} \code{hsa-miR-200a-3p} \code{hsa-miR-200a-5p} \code{hsa-miR-200b-3p} \code{hsa-miR-200b-5p} \code{hsa-miR-200c-3p} \code{hsa-miR-200c-5p} \code{hsa-miR-202-3p} \code{hsa-miR-202-5p} \code{hsa-miR-203} \code{hsa-miR-204-5p} \code{hsa-miR-205-3p} \code{hsa-miR-205-5p} \code{hsa-miR-2052} \code{hsa-miR-2053} \code{hsa-miR-206} \code{hsa-miR-208a} \code{hsa-miR-208b} \code{hsa-miR-20a-3p} \code{hsa-miR-20a-5p} \code{hsa-miR-20b-3p} \code{hsa-miR-20b-5p} \code{hsa-miR-21-3p} \code{hsa-miR-21-5p} \code{hsa-miR-210} \code{hsa-miR-211-5p} \code{hsa-miR-2110} \code{hsa-miR-2113} \code{hsa-miR-2114-3p} \code{hsa-miR-2114-5p} \code{hsa-miR-2115-3p} \code{hsa-miR-2115-5p} \code{hsa-miR-2116-3p} \code{hsa-miR-2116-5p} \code{hsa-miR-2117} \code{hsa-miR-212-3p} \code{hsa-miR-214-3p} \code{hsa-miR-214-5p} \code{hsa-miR-215} \code{hsa-miR-216a} \code{hsa-miR-216b} \code{hsa-miR-217} \code{hsa-miR-218-1-3p} \code{hsa-miR-218-2-3p} \code{hsa-miR-218-5p} \code{hsa-miR-219-1-3p} \code{hsa-miR-219-2-3p} \code{hsa-miR-219-5p} \code{hsa-miR-22-3p} \code{hsa-miR-22-5p} \code{hsa-miR-221-3p} \code{hsa-miR-221-5p} \code{hsa-miR-222-3p} \code{hsa-miR-222-5p} \code{hsa-miR-223-3p} \code{hsa-miR-223-5p} \code{hsa-miR-224-3p} \code{hsa-miR-224-5p} \code{hsa-miR-2276} \code{hsa-miR-2277-3p} \code{hsa-miR-2277-5p} \code{hsa-miR-2278} \code{hsa-miR-2355-3p} \code{hsa-miR-2355-5p} \code{hsa-miR-23a-3p} \code{hsa-miR-23a-5p} \code{hsa-miR-23b-3p} \code{hsa-miR-23b-5p} \code{hsa-miR-23c} \code{hsa-miR-24-1-5p} \code{hsa-miR-24-2-5p} \code{hsa-miR-24-3p} \code{hsa-miR-25-3p} \code{hsa-miR-25-5p} \code{hsa-miR-26a-1-3p} \code{hsa-miR-26a-2-3p} \code{hsa-miR-26a-5p} \code{hsa-miR-26b-3p} \code{hsa-miR-26b-5p} \code{hsa-miR-27a-3p} \code{hsa-miR-27a-5p} \code{hsa-miR-27b-3p} \code{hsa-miR-27b-5p} \code{hsa-miR-28-3p} \code{hsa-miR-28-5p} \code{hsa-miR-2861} \code{hsa-miR-2909} \code{hsa-miR-296-3p} \code{hsa-miR-296-5p} \code{hsa-miR-297} \code{hsa-miR-298} \code{hsa-miR-299-3p} \code{hsa-miR-299-5p} \code{hsa-miR-29a-3p} \code{hsa-miR-29a-5p} \code{hsa-miR-29b-1-5p} \code{hsa-miR-29b-2-5p} \code{hsa-miR-29b-3p} \code{hsa-miR-29c-3p} \code{hsa-miR-29c-5p} \code{hsa-miR-300} \code{hsa-miR-301a-3p} \code{hsa-miR-301b} \code{hsa-miR-302a-3p} \code{hsa-miR-302a-5p} \code{hsa-miR-302b-3p} \code{hsa-miR-302b-5p} \code{hsa-miR-302c-3p} \code{hsa-miR-302c-5p} \code{hsa-miR-302d-3p} \code{hsa-miR-302d-5p} \code{hsa-miR-302e} \code{hsa-miR-302f} \code{hsa-miR-3065-3p} \code{hsa-miR-3065-5p} \code{hsa-miR-3074-3p} \code{hsa-miR-30a-3p} \code{hsa-miR-30a-5p} \code{hsa-miR-30b-3p} \code{hsa-miR-30b-5p} \code{hsa-miR-30c-1-3p} \code{hsa-miR-30c-2-3p} \code{hsa-miR-30c-5p} \code{hsa-miR-30d-3p} \code{hsa-miR-30d-5p} \code{hsa-miR-30e-3p} \code{hsa-miR-30e-5p} \code{hsa-miR-31-3p} \code{hsa-miR-31-5p} \code{hsa-miR-3115} \code{hsa-miR-3116} \code{hsa-miR-3117-3p} \code{hsa-miR-3118} \code{hsa-miR-3119} \code{hsa-miR-3120-3p} \code{hsa-miR-3121-3p} \code{hsa-miR-3122} \code{hsa-miR-3123} \code{hsa-miR-3124-5p} \code{hsa-miR-3125} \code{hsa-miR-3126-3p} \code{hsa-miR-3126-5p} \code{hsa-miR-3127-5p} \code{hsa-miR-3128} \code{hsa-miR-3129-5p} \code{hsa-miR-3130-3p} \code{hsa-miR-3130-5p} \code{hsa-miR-3131} \code{hsa-miR-3132} \code{hsa-miR-3133} \code{hsa-miR-3134} \code{hsa-miR-3135a} \code{hsa-miR-3136-5p} \code{hsa-miR-3137} \code{hsa-miR-3138} \code{hsa-miR-3139} \code{hsa-miR-3140-3p} \code{hsa-miR-3141} \code{hsa-miR-3142} \code{hsa-miR-3143} \code{hsa-miR-3144-3p} \code{hsa-miR-3144-5p} \code{hsa-miR-3145-3p} \code{hsa-miR-3146} \code{hsa-miR-3147} \code{hsa-miR-3148} \code{hsa-miR-3149} \code{hsa-miR-3150a-3p} \code{hsa-miR-3150b-3p} \code{hsa-miR-3151} \code{hsa-miR-3152-3p} \code{hsa-miR-3153} \code{hsa-miR-3154} \code{hsa-miR-3155a} \code{hsa-miR-3156-5p} \code{hsa-miR-3157-5p} \code{hsa-miR-3158-3p} \code{hsa-miR-3159} \code{hsa-miR-3160-3p} \code{hsa-miR-3161} \code{hsa-miR-3162-5p} \code{hsa-miR-3163} \code{hsa-miR-3164} \code{hsa-miR-3165} \code{hsa-miR-3166} \code{hsa-miR-3167} \code{hsa-miR-3168} \code{hsa-miR-3169} \code{hsa-miR-3170} \code{hsa-miR-3171} \code{hsa-miR-3173-3p} \code{hsa-miR-3174} \code{hsa-miR-3175} \code{hsa-miR-3176} \code{hsa-miR-3177-3p} \code{hsa-miR-3178} \code{hsa-miR-3179} \code{hsa-miR-3180} \code{hsa-miR-3180-3p} \code{hsa-miR-3180-5p} \code{hsa-miR-3181} \code{hsa-miR-3182} \code{hsa-miR-3183} \code{hsa-miR-3184-5p} \code{hsa-miR-3185} \code{hsa-miR-3186-3p} \code{hsa-miR-3186-5p} \code{hsa-miR-3187-3p} \code{hsa-miR-3188} \code{hsa-miR-3189-3p} \code{hsa-miR-3190-3p} \code{hsa-miR-3191-3p} \code{hsa-miR-3192} \code{hsa-miR-3193} \code{hsa-miR-3194-5p} \code{hsa-miR-3195} \code{hsa-miR-3196} \code{hsa-miR-3197} \code{hsa-miR-3198} \code{hsa-miR-3199} \code{hsa-miR-32-3p} \code{hsa-miR-32-5p} \code{hsa-miR-3200-3p} \code{hsa-miR-3200-5p} \code{hsa-miR-3201} \code{hsa-miR-3202} \code{hsa-miR-320a} \code{hsa-miR-320b} \code{hsa-miR-320c} \code{hsa-miR-320d} \code{hsa-miR-320e} \code{hsa-miR-323a-3p} \code{hsa-miR-323a-5p} \code{hsa-miR-323b-3p} \code{hsa-miR-323b-5p} \code{hsa-miR-324-3p} \code{hsa-miR-324-5p} \code{hsa-miR-325} \code{hsa-miR-326} \code{hsa-miR-328} \code{hsa-miR-329} \code{hsa-miR-330-3p} \code{hsa-miR-330-5p} \code{hsa-miR-331-3p} \code{hsa-miR-331-5p} \code{hsa-miR-335-3p} \code{hsa-miR-335-5p} \code{hsa-miR-337-3p} \code{hsa-miR-337-5p} \code{hsa-miR-338-3p} \code{hsa-miR-338-5p} \code{hsa-miR-339-3p} \code{hsa-miR-339-5p} \code{hsa-miR-33a-3p} \code{hsa-miR-33a-5p} \code{hsa-miR-33b-3p} \code{hsa-miR-33b-5p} \code{hsa-miR-340-3p} \code{hsa-miR-340-5p} \code{hsa-miR-342-3p} \code{hsa-miR-342-5p} \code{hsa-miR-345-5p} \code{hsa-miR-346} \code{hsa-miR-34a-3p} \code{hsa-miR-34a-5p} \code{hsa-miR-34b-3p} \code{hsa-miR-34b-5p} \code{hsa-miR-34c-3p} \code{hsa-miR-34c-5p} \code{hsa-miR-3605-3p} \code{hsa-miR-3605-5p} \code{hsa-miR-3606} \code{hsa-miR-3607-3p} \code{hsa-miR-3607-5p} \code{hsa-miR-3609} \code{hsa-miR-361-3p} \code{hsa-miR-361-5p} \code{hsa-miR-3610} \code{hsa-miR-3611} \code{hsa-miR-3612} \code{hsa-miR-3613-3p} \code{hsa-miR-3613-5p} \code{hsa-miR-3614-3p} \code{hsa-miR-3614-5p} \code{hsa-miR-3615} \code{hsa-miR-3616-3p} \code{hsa-miR-3616-5p} \code{hsa-miR-3617} \code{hsa-miR-3618} \code{hsa-miR-3619-5p} \code{hsa-miR-362-3p} \code{hsa-miR-362-5p} \code{hsa-miR-3620} \code{hsa-miR-3621} \code{hsa-miR-3622a-3p} \code{hsa-miR-3622a-5p} \code{hsa-miR-3622b-3p} \code{hsa-miR-3622b-5p} \code{hsa-miR-363-3p} \code{hsa-miR-363-5p} \code{hsa-miR-3646} \code{hsa-miR-3647-3p_v17.0} \code{hsa-miR-3647-5p_v17.0} \code{hsa-miR-3648} \code{hsa-miR-3649} \code{hsa-miR-3650} \code{hsa-miR-3651} \code{hsa-miR-3652} \code{hsa-miR-3653} \code{hsa-miR-3654} \code{hsa-miR-3655} \code{hsa-miR-3656} \code{hsa-miR-3657} \code{hsa-miR-3658} \code{hsa-miR-3659} \code{hsa-miR-365a-3p} \code{hsa-miR-365b-5p} \code{hsa-miR-3660} \code{hsa-miR-3661} \code{hsa-miR-3662} \code{hsa-miR-3663-3p} \code{hsa-miR-3663-5p} \code{hsa-miR-3664-5p} \code{hsa-miR-3665} \code{hsa-miR-3666} \code{hsa-miR-3667-3p} \code{hsa-miR-3667-5p} \code{hsa-miR-3668} \code{hsa-miR-3669} \code{hsa-miR-367-3p} \code{hsa-miR-367-5p} \code{hsa-miR-3670} \code{hsa-miR-3671} \code{hsa-miR-3672} \code{hsa-miR-3673} \code{hsa-miR-3674} \code{hsa-miR-3675-3p} \code{hsa-miR-3675-5p} \code{hsa-miR-3676-3p} \code{hsa-miR-3677-3p} \code{hsa-miR-3678-3p} \code{hsa-miR-3678-5p} \code{hsa-miR-3679-3p} \code{hsa-miR-3679-5p} \code{hsa-miR-3680-3p} \code{hsa-miR-3680-5p} \code{hsa-miR-3681-3p} \code{hsa-miR-3681-5p} \code{hsa-miR-3682-3p} \code{hsa-miR-3683} \code{hsa-miR-3684} \code{hsa-miR-3685} \code{hsa-miR-3686} \code{hsa-miR-3687} \code{hsa-miR-3688-3p} \code{hsa-miR-3689a-3p} \code{hsa-miR-3689a-5p} \code{hsa-miR-3689b-3p} \code{hsa-miR-369-3p} \code{hsa-miR-369-5p} \code{hsa-miR-3690} \code{hsa-miR-3691-5p} \code{hsa-miR-3692-3p} \code{hsa-miR-3692-5p} \code{hsa-miR-370} \code{hsa-miR-3713} \code{hsa-miR-3714} \code{hsa-miR-371a-3p} \code{hsa-miR-371a-5p} \code{hsa-miR-372} \code{hsa-miR-373-3p} \code{hsa-miR-373-5p} \code{hsa-miR-374a-3p} \code{hsa-miR-374a-5p} \code{hsa-miR-374b-3p} \code{hsa-miR-374b-5p} \code{hsa-miR-374c-5p} \code{hsa-miR-375} \code{hsa-miR-376a-3p} \code{hsa-miR-376a-5p} \code{hsa-miR-376b} \code{hsa-miR-376c} \code{hsa-miR-377-3p} \code{hsa-miR-377-5p} \code{hsa-miR-378_v17.0} \code{hsa-miR-378a-5p} \code{hsa-miR-378b} \code{hsa-miR-378c} \code{hsa-miR-379-3p} \code{hsa-miR-379-5p} \code{hsa-miR-380-3p} \code{hsa-miR-380-5p} \code{hsa-miR-381} \code{hsa-miR-382-5p} \code{hsa-miR-383} \code{hsa-miR-384} \code{hsa-miR-3907} \code{hsa-miR-3908} \code{hsa-miR-3909} \code{hsa-miR-3910} \code{hsa-miR-3911} \code{hsa-miR-3912} \code{hsa-miR-3913-5p} \code{hsa-miR-3914} \code{hsa-miR-3915} \code{hsa-miR-3916} \code{hsa-miR-3917} \code{hsa-miR-3918} \code{hsa-miR-3919} \code{hsa-miR-3920} \code{hsa-miR-3921} \code{hsa-miR-3922-3p} \code{hsa-miR-3923} \code{hsa-miR-3924} \code{hsa-miR-3925-5p} \code{hsa-miR-3926} \code{hsa-miR-3927} \code{hsa-miR-3928} \code{hsa-miR-3929} \code{hsa-miR-3934} \code{hsa-miR-3935} \code{hsa-miR-3936} \code{hsa-miR-3937} \code{hsa-miR-3938} \code{hsa-miR-3939} \code{hsa-miR-3940-3p} \code{hsa-miR-3941} \code{hsa-miR-3942-5p} \code{hsa-miR-3943} \code{hsa-miR-3944-3p} \code{hsa-miR-3945} \code{hsa-miR-409-3p} \code{hsa-miR-409-5p} \code{hsa-miR-410} \code{hsa-miR-411-3p} \code{hsa-miR-411-5p} \code{hsa-miR-412} \code{hsa-miR-421} \code{hsa-miR-422a} \code{hsa-miR-423-3p} \code{hsa-miR-423-5p} \code{hsa-miR-424-3p} \code{hsa-miR-424-5p} \code{hsa-miR-425-3p} \code{hsa-miR-425-5p} \code{hsa-miR-4251} \code{hsa-miR-4252} \code{hsa-miR-4253} \code{hsa-miR-4254} \code{hsa-miR-4255} \code{hsa-miR-4256} \code{hsa-miR-4257} \code{hsa-miR-4258} \code{hsa-miR-4259} \code{hsa-miR-4260} \code{hsa-miR-4261} \code{hsa-miR-4262} \code{hsa-miR-4263} \code{hsa-miR-4264} \code{hsa-miR-4265} \code{hsa-miR-4266} \code{hsa-miR-4267} \code{hsa-miR-4268} \code{hsa-miR-4269} \code{hsa-miR-4270} \code{hsa-miR-4271} \code{hsa-miR-4272} \code{hsa-miR-4273} \code{hsa-miR-4274} \code{hsa-miR-4275} \code{hsa-miR-4276} \code{hsa-miR-4277} \code{hsa-miR-4278} \code{hsa-miR-4279} \code{hsa-miR-4280} \code{hsa-miR-4281} \code{hsa-miR-4282} \code{hsa-miR-4283} \code{hsa-miR-4284} \code{hsa-miR-4285} \code{hsa-miR-4286} \code{hsa-miR-4287} \code{hsa-miR-4288} \code{hsa-miR-4289} \code{hsa-miR-429} \code{hsa-miR-4290} \code{hsa-miR-4291} \code{hsa-miR-4292} \code{hsa-miR-4293} \code{hsa-miR-4294} \code{hsa-miR-4295} \code{hsa-miR-4296} \code{hsa-miR-4297} \code{hsa-miR-4298} \code{hsa-miR-4299} \code{hsa-miR-4300} \code{hsa-miR-4301} \code{hsa-miR-4302} \code{hsa-miR-4303} \code{hsa-miR-4304} \code{hsa-miR-4305} \code{hsa-miR-4306} \code{hsa-miR-4307} \code{hsa-miR-4308} \code{hsa-miR-4309} \code{hsa-miR-431-3p} \code{hsa-miR-431-5p} \code{hsa-miR-4310} \code{hsa-miR-4311} \code{hsa-miR-4312} \code{hsa-miR-4313} \code{hsa-miR-4314} \code{hsa-miR-4315} \code{hsa-miR-4316} \code{hsa-miR-4317} \code{hsa-miR-4318} \code{hsa-miR-4319} \code{hsa-miR-432-3p} \code{hsa-miR-432-5p} \code{hsa-miR-4320} \code{hsa-miR-4321} \code{hsa-miR-4322} \code{hsa-miR-4323} \code{hsa-miR-4324} \code{hsa-miR-4325} \code{hsa-miR-4326} \code{hsa-miR-4327} \code{hsa-miR-4328} \code{hsa-miR-4329} \code{hsa-miR-433} \code{hsa-miR-4330} \code{hsa-miR-448} \code{hsa-miR-449a} \code{hsa-miR-449b-3p} \code{hsa-miR-449b-5p} \code{hsa-miR-449c-3p} \code{hsa-miR-449c-5p} \code{hsa-miR-450a-5p} \code{hsa-miR-450b-3p} \code{hsa-miR-450b-5p} \code{hsa-miR-451a} \code{hsa-miR-452-3p} \code{hsa-miR-452-5p} \code{hsa-miR-454-3p} \code{hsa-miR-454-5p} \code{hsa-miR-455-3p} \code{hsa-miR-455-5p} \code{hsa-miR-466} \code{hsa-miR-483-3p} \code{hsa-miR-483-5p} \code{hsa-miR-484} \code{hsa-miR-485-3p} \code{hsa-miR-485-5p} \code{hsa-miR-486-3p} \code{hsa-miR-486-5p} \code{hsa-miR-487a} \code{hsa-miR-487b} \code{hsa-miR-488-3p} \code{hsa-miR-488-5p} \code{hsa-miR-489} \code{hsa-miR-490-3p} \code{hsa-miR-490-5p} \code{hsa-miR-491-3p} \code{hsa-miR-491-5p} \code{hsa-miR-492} \code{hsa-miR-493-3p} \code{hsa-miR-493-5p} \code{hsa-miR-494} \code{hsa-miR-495} \code{hsa-miR-496} \code{hsa-miR-497-3p} \code{hsa-miR-497-5p} \code{hsa-miR-498} \code{hsa-miR-499a-3p} \code{hsa-miR-499a-5p} \code{hsa-miR-500a-3p} \code{hsa-miR-500a-5p} \code{hsa-miR-500b} \code{hsa-miR-501-3p} \code{hsa-miR-501-5p} \code{hsa-miR-502-3p} \code{hsa-miR-502-5p} \code{hsa-miR-503} \code{hsa-miR-504} \code{hsa-miR-505-3p} \code{hsa-miR-505-5p} \code{hsa-miR-506-3p} \code{hsa-miR-507} \code{hsa-miR-508-3p} \code{hsa-miR-508-5p} \code{hsa-miR-509-3-5p} \code{hsa-miR-509-3p} \code{hsa-miR-509-5p} \code{hsa-miR-510} \code{hsa-miR-511} \code{hsa-miR-512-3p} \code{hsa-miR-512-5p} \code{hsa-miR-513a-3p} \code{hsa-miR-513a-5p} \code{hsa-miR-513b} \code{hsa-miR-513c-5p} \code{hsa-miR-514a-3p} \code{hsa-miR-514b-3p} \code{hsa-miR-514b-5p} \code{hsa-miR-515-3p} \code{hsa-miR-515-5p} \code{hsa-miR-516a-3p} \code{hsa-miR-516a-5p} \code{hsa-miR-516b-5p} \code{hsa-miR-517-5p} \code{hsa-miR-517a-3p} \code{hsa-miR-517b_v17.0} \code{hsa-miR-517c-3p} \code{hsa-miR-518a-3p} \code{hsa-miR-518a-5p} \code{hsa-miR-518b} \code{hsa-miR-518c-3p} \code{hsa-miR-518c-5p} \code{hsa-miR-518d-3p} \code{hsa-miR-518e-3p} \code{hsa-miR-518e-5p} \code{hsa-miR-518f-3p} \code{hsa-miR-518f-5p} \code{hsa-miR-519b-3p} \code{hsa-miR-519c-3p} \code{hsa-miR-519d} \code{hsa-miR-519e-3p} \code{hsa-miR-519e-5p} \code{hsa-miR-520a-3p} \code{hsa-miR-520a-5p} \code{hsa-miR-520b} \code{hsa-miR-520c-3p} \code{hsa-miR-520d-3p} \code{hsa-miR-520e} \code{hsa-miR-520f} \code{hsa-miR-520g} \code{hsa-miR-520h} \code{hsa-miR-521} \code{hsa-miR-522-3p} \code{hsa-miR-523-3p} \code{hsa-miR-524-3p} \code{hsa-miR-525-3p} \code{hsa-miR-525-5p} \code{hsa-miR-526b-3p} \code{hsa-miR-526b-5p} \code{hsa-miR-532-3p} \code{hsa-miR-532-5p} \code{hsa-miR-539-5p} \code{hsa-miR-541-3p} \code{hsa-miR-541-5p} \code{hsa-miR-542-3p} \code{hsa-miR-542-5p} \code{hsa-miR-543} \code{hsa-miR-544a} \code{hsa-miR-544b} \code{hsa-miR-545-3p} \code{hsa-miR-545-5p} \code{hsa-miR-548a-3p} \code{hsa-miR-548a-5p} \code{hsa-miR-548aa} \code{hsa-miR-548am-5p} \code{hsa-miR-548av-3p} \code{hsa-miR-548b-3p} \code{hsa-miR-548b-5p} \code{hsa-miR-548c-3p} \code{hsa-miR-548d-3p} \code{hsa-miR-548d-5p} \code{hsa-miR-548e} \code{hsa-miR-548f} \code{hsa-miR-548g-3p} \code{hsa-miR-548h-3p} \code{hsa-miR-548h-5p} \code{hsa-miR-548i} \code{hsa-miR-548j} \code{hsa-miR-548k} \code{hsa-miR-548l} \code{hsa-miR-548m} \code{hsa-miR-548n} \code{hsa-miR-548p} \code{hsa-miR-548q} \code{hsa-miR-548s} \code{hsa-miR-548t-5p} \code{hsa-miR-548u} \code{hsa-miR-548v} \code{hsa-miR-548w} \code{hsa-miR-548x_v16.0} \code{hsa-miR-548y} \code{hsa-miR-549} \code{hsa-miR-550a-3p} \code{hsa-miR-550a-5p} \code{hsa-miR-550b-3p} \code{hsa-miR-551a} \code{hsa-miR-551b-3p} \code{hsa-miR-551b-5p} \code{hsa-miR-552} \code{hsa-miR-553} \code{hsa-miR-554} \code{hsa-miR-555} \code{hsa-miR-556-3p} \code{hsa-miR-556-5p} \code{hsa-miR-557} \code{hsa-miR-558} \code{hsa-miR-559} \code{hsa-miR-561-3p} \code{hsa-miR-562} \code{hsa-miR-563} \code{hsa-miR-564} \code{hsa-miR-566} \code{hsa-miR-567} \code{hsa-miR-568} \code{hsa-miR-569} \code{hsa-miR-570-3p} \code{hsa-miR-571} \code{hsa-miR-572} \code{hsa-miR-573} \code{hsa-miR-574-3p} \code{hsa-miR-574-5p} \code{hsa-miR-575} \code{hsa-miR-576-3p} \code{hsa-miR-576-5p} \code{hsa-miR-577} \code{hsa-miR-578} \code{hsa-miR-579} \code{hsa-miR-580} \code{hsa-miR-581} \code{hsa-miR-582-3p} \code{hsa-miR-582-5p} \code{hsa-miR-583} \code{hsa-miR-584-5p} \code{hsa-miR-585} \code{hsa-miR-586} \code{hsa-miR-587} \code{hsa-miR-588} \code{hsa-miR-589-3p} \code{hsa-miR-589-5p} \code{hsa-miR-590-3p} \code{hsa-miR-590-5p} \code{hsa-miR-591} \code{hsa-miR-592} \code{hsa-miR-593-3p} \code{hsa-miR-593-5p} \code{hsa-miR-595} \code{hsa-miR-596} \code{hsa-miR-597} \code{hsa-miR-598} \code{hsa-miR-599} \code{hsa-miR-600} \code{hsa-miR-601} \code{hsa-miR-602} \code{hsa-miR-603} \code{hsa-miR-604} \code{hsa-miR-605} \code{hsa-miR-606} \code{hsa-miR-607} \code{hsa-miR-608} \code{hsa-miR-609} \code{hsa-miR-610} \code{hsa-miR-611} \code{hsa-miR-612} \code{hsa-miR-613} \code{hsa-miR-614} \code{hsa-miR-615-3p} \code{hsa-miR-615-5p} \code{hsa-miR-616-3p} \code{hsa-miR-616-5p} \code{hsa-miR-617} \code{hsa-miR-618} \code{hsa-miR-619} \code{hsa-miR-620} \code{hsa-miR-621} \code{hsa-miR-622} \code{hsa-miR-623} \code{hsa-miR-624-3p} \code{hsa-miR-624-5p} \code{hsa-miR-625-3p} \code{hsa-miR-625-5p} \code{hsa-miR-626} \code{hsa-miR-627} \code{hsa-miR-628-3p} \code{hsa-miR-628-5p} \code{hsa-miR-629-3p} \code{hsa-miR-629-5p} \code{hsa-miR-630} \code{hsa-miR-631} \code{hsa-miR-632} \code{hsa-miR-633} \code{hsa-miR-634} \code{hsa-miR-635} \code{hsa-miR-636} \code{hsa-miR-637} \code{hsa-miR-638} \code{hsa-miR-639} \code{hsa-miR-640} \code{hsa-miR-641} \code{hsa-miR-642a-5p} \code{hsa-miR-642b-3p} \code{hsa-miR-643} \code{hsa-miR-644a} \code{hsa-miR-645} \code{hsa-miR-646} \code{hsa-miR-647} \code{hsa-miR-648} \code{hsa-miR-649} \code{hsa-miR-650} \code{hsa-miR-651} \code{hsa-miR-652-3p} \code{hsa-miR-653} \code{hsa-miR-654-3p} \code{hsa-miR-654-5p} \code{hsa-miR-655} \code{hsa-miR-656} \code{hsa-miR-657} \code{hsa-miR-658} \code{hsa-miR-659-3p} \code{hsa-miR-660-5p} \code{hsa-miR-661} \code{hsa-miR-662} \code{hsa-miR-663a} \code{hsa-miR-663b} \code{hsa-miR-664-3p} \code{hsa-miR-664-5p} \code{hsa-miR-665} \code{hsa-miR-668} \code{hsa-miR-670} \code{hsa-miR-671-3p} \code{hsa-miR-671-5p} \code{hsa-miR-675-3p} \code{hsa-miR-675-5p} \code{hsa-miR-676-3p} \code{hsa-miR-676-5p} \code{hsa-miR-7-1-3p} \code{hsa-miR-7-2-3p} \code{hsa-miR-7-5p} \code{hsa-miR-708-3p} \code{hsa-miR-708-5p} \code{hsa-miR-711} \code{hsa-miR-718} \code{hsa-miR-720} \code{hsa-miR-744-3p} \code{hsa-miR-744-5p} \code{hsa-miR-758} \code{hsa-miR-759} \code{hsa-miR-760} \code{hsa-miR-761} \code{hsa-miR-762} \code{hsa-miR-764} \code{hsa-miR-765} \code{hsa-miR-766-3p} \code{hsa-miR-767-3p} \code{hsa-miR-767-5p} \code{hsa-miR-769-3p} \code{hsa-miR-769-5p} \code{hsa-miR-770-5p} \code{hsa-miR-802} \code{hsa-miR-873-5p} \code{hsa-miR-874} \code{hsa-miR-875-3p} \code{hsa-miR-875-5p} \code{hsa-miR-876-3p} \code{hsa-miR-876-5p} \code{hsa-miR-877-3p} \code{hsa-miR-877-5p} \code{hsa-miR-885-3p} \code{hsa-miR-885-5p} \code{hsa-miR-887} \code{hsa-miR-888-3p} \code{hsa-miR-888-5p} \code{hsa-miR-889} \code{hsa-miR-890} \code{hsa-miR-891a} \code{hsa-miR-891b} \code{hsa-miR-892a} \code{hsa-miR-892b} \code{hsa-miR-9-3p} \code{hsa-miR-9-5p} \code{hsa-miR-920} \code{hsa-miR-921} \code{hsa-miR-922} \code{hsa-miR-924} \code{hsa-miR-92a-1-5p} \code{hsa-miR-92a-2-5p} \code{hsa-miR-92a-3p} \code{hsa-miR-92b-3p} \code{hsa-miR-92b-5p} \code{hsa-miR-93-3p} \code{hsa-miR-93-5p} \code{hsa-miR-933} \code{hsa-miR-934} \code{hsa-miR-935} \code{hsa-miR-936} \code{hsa-miR-937} \code{hsa-miR-938} \code{hsa-miR-939} \code{hsa-miR-940} \code{hsa-miR-941} \code{hsa-miR-942} \code{hsa-miR-943} \code{hsa-miR-944} \code{hsa-miR-95} \code{hsa-miR-96-3p} \code{hsa-miR-96-5p} \code{hsa-miR-98} \code{hsa-miR-99a-3p} \code{hsa-miR-99a-5p} \code{hsa-miR-99b-3p} \code{hsa-miR-99b-5p} \code{hsv1-miR-H1} \code{hsv1-miR-H1*} \code{hsv1-miR-H11} \code{hsv1-miR-H12} \code{hsv1-miR-H13} \code{hsv1-miR-H14-3p} \code{hsv1-miR-H14-5p} \code{hsv1-miR-H15} \code{hsv1-miR-H16} \code{hsv1-miR-H17} \code{hsv1-miR-H18} \code{hsv1-miR-H2} \code{hsv1-miR-H2*} \code{hsv1-miR-H3} \code{hsv1-miR-H3*} \code{hsv1-miR-H4-3p} \code{hsv1-miR-H4-5p} \code{hsv1-miR-H5-3p} \code{hsv1-miR-H5-5p} \code{hsv1-miR-H6-3p} \code{hsv1-miR-H6-5p} \code{hsv1-miR-H7} \code{hsv1-miR-H7*} \code{hsv1-miR-H8} \code{hsv1-miR-H8*} \code{hsv2-miR-H10} \code{hsv2-miR-H11*} \code{hsv2-miR-H13} \code{hsv2-miR-H19} \code{hsv2-miR-H2} \code{hsv2-miR-H20} \code{hsv2-miR-H21} \code{hsv2-miR-H22} \code{hsv2-miR-H23} \code{hsv2-miR-H23*} \code{hsv2-miR-H24} \code{hsv2-miR-H25} \code{hsv2-miR-H3} \code{hsv2-miR-H4-3p} \code{hsv2-miR-H4-5p} \code{hsv2-miR-H5} \code{hsv2-miR-H6} \code{hsv2-miR-H6*} \code{hsv2-miR-H7-3p} \code{hsv2-miR-H7-5p} \code{hsv2-miR-H9-3p} \code{hsv2-miR-H9-5p} \code{hur_1} \code{hur_2} \code{hur_4} \code{hur_5} \code{hur_6} \code{jcv-miR-J1-5p} \code{kshv-miR-K12-1} \code{kshv-miR-K12-1*} \code{kshv-miR-K12-10a*} \code{kshv-miR-K12-10b} \code{kshv-miR-K12-11} \code{kshv-miR-K12-11*} \code{kshv-miR-K12-12} \code{kshv-miR-K12-12*} \code{kshv-miR-K12-2} \code{kshv-miR-K12-2*} \code{kshv-miR-K12-3} \code{kshv-miR-K12-3*} \code{kshv-miR-K12-4-3p} \code{kshv-miR-K12-4-5p} \code{kshv-miR-K12-5} \code{kshv-miR-K12-5*} \code{kshv-miR-K12-6-3p} \code{kshv-miR-K12-6-5p} \code{kshv-miR-K12-7} \code{kshv-miR-K12-7*} \code{kshv-miR-K12-8} \code{kshv-miR-K12-8*} \code{kshv-miR-K12-9} \code{kshv-miR-K12-9*} \code{miRNABrightCorner30} \code{mr_1} \code{NC1_00000197} \code{NC1_00000215} \code{NC2_00079215} \code{NC2_00092197} \code{NC2_00106057} \code{NC2_00122731} \code{NegativeControl}}
    \item{\code{ex1_02}}{a numeric vector}
    \item{\code{ex1_03}}{a numeric vector}
    \item{\code{ex1_05}}{a numeric vector}
    \item{\code{ex1_06}}{a numeric vector}
    \item{\code{ex1_07}}{a numeric vector}
    \item{\code{ex1_08}}{a numeric vector}
    \item{\code{ex1_09}}{a numeric vector}
    \item{\code{ex1_10}}{a numeric vector}
    \item{\code{ex1_12}}{a numeric vector}
    \item{\code{ex1_13}}{a numeric vector}
    \item{\code{ex1_14}}{a numeric vector}
    \item{\code{ex1_26}}{a numeric vector}
    \item{\code{ex1_27}}{a numeric vector}
    \item{\code{ex1_28}}{a numeric vector}
    \item{\code{ex1_39}}{a numeric vector}
    \item{\code{ex1_40}}{a numeric vector}
    \item{\code{ex2_01}}{a numeric vector}
    \item{\code{ex2_04}}{a numeric vector}
    \item{\code{ex2_15}}{a numeric vector}
    \item{\code{ex2_16}}{a numeric vector}
    \item{\code{ex2_17}}{a numeric vector}
    \item{\code{ex2_18}}{a numeric vector}
    \item{\code{ex2_19}}{a numeric vector}
    \item{\code{ex2_20}}{a numeric vector}
    \item{\code{ex2_21}}{a numeric vector}
    \item{\code{ex2_22}}{a numeric vector}
    \item{\code{ex2_23}}{a numeric vector}
    \item{\code{ex2_30}}{a numeric vector}
    \item{\code{ex2_31}}{a numeric vector}
    \item{\code{ex2_32}}{a numeric vector}
    \item{\code{ex2_33}}{a numeric vector}
    \item{\code{ex2_34}}{a numeric vector}
    \item{\code{ex2_35}}{a numeric vector}
    \item{\code{ex2_37}}{a numeric vector}
    \item{\code{ex2_38}}{a numeric vector}
    \item{\code{ex2_41}}{a numeric vector}
    \item{\code{ex2_43}}{a numeric vector}
    \item{\code{ex2_44}}{a numeric vector}
    \item{\code{ex2_46}}{a numeric vector}
    \item{\code{ex2_47}}{a numeric vector}
    \item{\code{ex2_48}}{a numeric vector}
    \item{\code{ex2_49}}{a numeric vector}
    \item{\code{ex2_51}}{a numeric vector}
    \item{\code{ex2_52}}{a numeric vector}
    \item{\code{ex2_53}}{a numeric vector}
    \item{\code{ex2_55}}{a numeric vector}
    \item{\code{ex2_56}}{a numeric vector}
    \item{\code{ex2_57}}{a numeric vector}
    \item{\code{ex2_58}}{a numeric vector}
    \item{\code{ex2_59}}{a numeric vector}
  }
}
\details{
This a micRNA expression dataset for validation of PanNET subtypes. micRNAs are on the rows and the first column contain the micRNAs names.
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(PanNETsubtypesmicRNAVal.Indata)
## maybe str(PanNETsubtypesmicRNAVal.Indata) ; plot(PanNETsubtypesmicRNAVal.Indata) ...
}
\keyword{datasets}
