\name{PanNETsubtypesmicRNAVal.IndSAMdata}
\alias{PanNETsubtypesmicRNAVal.IndSAMdata}
\docType{data}
\title{
PanNETassigner micRNA expression signature.
}
\description{
This dataset contain micRNA signature for identifying the PanNET subtypes.

}
\usage{data("PanNETsubtypesmicRNAVal.IndSAMdata")}
\format{
  A data frame with 41 observations on the following variable.
  \describe{
    \item{\code{ProbeID}}{a character vector}
  }
}
\details{
This dataset contain micRNA signature for identifying the PanNET subtypes.
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(PanNETsubtypesmicRNAVal.IndSAMdata)
## maybe str(PanNETsubtypesmicRNAVal.IndSAMdata) ; plot(PanNETsubtypesmicRNAVal.IndSAMdata) ...
}
\keyword{datasets}
