\name{PanNETsubtypesmRNAVal.IndSAMdata}
\alias{PanNETsubtypesmRNAVal.IndSAMdata}
\docType{data}
\title{ PanNETassigner gene expression signature.
}
\description{
This dataset contain 221 PanNETassigner gene signature for identifying the PanNET subtypes.
}
\usage{data("PanNETsubtypesmRNAVal.IndSAMdata")}
\format{
  A data frame with 221 observations on the following 87 variables.
  \describe{
    \item{\code{X}}{a character vector}
    \item{\code{X90}}{a numeric vector}
    \item{\code{X86}}{a numeric vector}
    \item{\code{X34}}{a numeric vector}
    \item{\code{X120}}{a numeric vector}
    \item{\code{X114}}{a numeric vector}
    \item{\code{X93}}{a numeric vector}
    \item{\code{X84}}{a numeric vector}
    \item{\code{X74}}{a numeric vector}
    \item{\code{X69}}{a numeric vector}
    \item{\code{X57}}{a numeric vector}
    \item{\code{X56}}{a numeric vector}
    \item{\code{X51}}{a numeric vector}
    \item{\code{X44}}{a numeric vector}
    \item{\code{X40}}{a numeric vector}
    \item{\code{X39}}{a numeric vector}
    \item{\code{X38}}{a numeric vector}
    \item{\code{X55}}{a numeric vector}
    \item{\code{X112}}{a numeric vector}
    \item{\code{X138}}{a numeric vector}
    \item{\code{X102}}{a numeric vector}
    \item{\code{X54}}{a numeric vector}
    \item{\code{X96}}{a numeric vector}
    \item{\code{X52}}{a numeric vector}
    \item{\code{X117}}{a numeric vector}
    \item{\code{X60}}{a numeric vector}
    \item{\code{X118}}{a numeric vector}
    \item{\code{X100}}{a numeric vector}
    \item{\code{X94}}{a numeric vector}
    \item{\code{X92}}{a numeric vector}
    \item{\code{X47}}{a numeric vector}
    \item{\code{X127}}{a numeric vector}
    \item{\code{X61}}{a numeric vector}
    \item{\code{X59}}{a numeric vector}
    \item{\code{X42}}{a numeric vector}
    \item{\code{X35}}{a numeric vector}
    \item{\code{X41}}{a numeric vector}
    \item{\code{X62}}{a numeric vector}
    \item{\code{X133}}{a numeric vector}
    \item{\code{X126}}{a numeric vector}
    \item{\code{X75}}{a numeric vector}
    \item{\code{X36}}{a numeric vector}
    \item{\code{X119}}{a numeric vector}
    \item{\code{X135}}{a numeric vector}
    \item{\code{X132}}{a numeric vector}
    \item{\code{X137}}{a numeric vector}
    \item{\code{X131}}{a numeric vector}
    \item{\code{X134}}{a numeric vector}
    \item{\code{X50}}{a numeric vector}
    \item{\code{X46}}{a numeric vector}
    \item{\code{X65}}{a numeric vector}
    \item{\code{X128}}{a numeric vector}
    \item{\code{X97}}{a numeric vector}
    \item{\code{X95}}{a numeric vector}
    \item{\code{X89}}{a numeric vector}
    \item{\code{X67}}{a numeric vector}
    \item{\code{X48}}{a numeric vector}
    \item{\code{X37}}{a numeric vector}
    \item{\code{X129}}{a numeric vector}
    \item{\code{X115}}{a numeric vector}
    \item{\code{X99}}{a numeric vector}
    \item{\code{X64}}{a numeric vector}
    \item{\code{X33}}{a numeric vector}
    \item{\code{X130}}{a numeric vector}
    \item{\code{X124}}{a numeric vector}
    \item{\code{X116}}{a numeric vector}
    \item{\code{X98}}{a numeric vector}
    \item{\code{X73}}{a numeric vector}
    \item{\code{X72}}{a numeric vector}
    \item{\code{X68}}{a numeric vector}
    \item{\code{X58}}{a numeric vector}
    \item{\code{X136}}{a numeric vector}
    \item{\code{X49}}{a numeric vector}
    \item{\code{X70}}{a numeric vector}
    \item{\code{X66}}{a numeric vector}
    \item{\code{X110}}{a numeric vector}
    \item{\code{X109}}{a numeric vector}
    \item{\code{X108}}{a numeric vector}
    \item{\code{X101}}{a numeric vector}
    \item{\code{X63}}{a numeric vector}
    \item{\code{X53}}{a numeric vector}
    \item{\code{X71}}{a numeric vector}
    \item{\code{X125}}{a numeric vector}
    \item{\code{X122}}{a numeric vector}
    \item{\code{X121}}{a numeric vector}
    \item{\code{X123}}{a numeric vector}
    \item{\code{X87}}{a numeric vector}
  }
}
\details{
This dataset contain 221 PanNETassigner gene signature for identifying the PanNET subtypes.
}
\source{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\references{
Anguraj Sadanandam, et all (2015). A cross-species analysis in pancreatic neuroendocrine tumors reveals molecular subtypes with distinctive clinical, metastatic, developmental, and metabolic characteristics. Cancer Discovery.
}
\examples{
data(PanNETsubtypesmRNAVal.IndSAMdata)
## maybe str(PanNETsubtypesmRNAVal.IndSAMdata) ; plot(PanNETsubtypesmRNAVal.IndSAMdata) ...
}
\keyword{datasets}
